Basic generation and writing API
