+++ renderite/generator/RenderiteGenerator/Generator/LegacyGenerator.cs
@@ -1,0 +1,623 @@
+﻿using System.Collections;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Reflection;
+using System.Runtime.InteropServices;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+using Mono.Cecil.Rocks;
+using RenderiteGenerator.Extensions;
+using RenderiteGenerator.Generator.Info;
+using RenderiteGenerator.Options;
+using TypeAttributes = System.Reflection.TypeAttributes;
+
+namespace RenderiteGenerator.Generator;
+
+public class LegacyGenerator : IDisposable
+{
+    private readonly Stream _fileStream;
+    private readonly StreamWriter _writer;
+    private readonly Assembly _assembly;
+    private readonly AssemblyDefinition _cecilAssembly;
+    private readonly bool _verbose;
+    private readonly bool _ilVerbose;
+    private readonly string _engineVersion;
+
+    private readonly List<Type> _generatedTypes = [];
+    private readonly Queue<Type> _typesToGenerate = [];
+
+    private readonly Type[] _types;
+    private readonly Type _iMemoryPackable;
+    
+    public LegacyGenerator(GeneratorOptions options)
+    {
+        if(File.Exists(options.OutputZigFile))
+            File.Delete(options.OutputZigFile);
+        
+        this._assembly = Assembly.LoadFrom(options.AssemblyPath);
+        this._cecilAssembly = AssemblyDefinition.ReadAssembly(options.AssemblyPath);
+        this._fileStream = File.OpenWrite(options.OutputZigFile);
+        this._writer = new StreamWriter(this._fileStream);
+        this._verbose = options.Verbose;
+        this._ilVerbose = options.IlVerbose;
+
+        this._engineVersion = "Unknown";
+        try
+        {
+            Assembly frooxEngine =
+                Assembly.LoadFrom(Path.Combine(Path.GetDirectoryName(options.AssemblyPath)!, "FrooxEngine.dll"));
+
+            this._engineVersion = frooxEngine.FullName!;
+        }
+        catch(Exception e)
+        {
+            Console.WriteLine($"Couldn't get FrooxEngine assembly to determine version: {e}");
+        }
+
+        this._types = this._assembly.GetTypes();
+        this._iMemoryPackable = this._types.First(t => t.Name == "IMemoryPackable");
+        
+        Console.WriteLine($"Generating for {this._engineVersion}");
+    }
+
+    public void Run()
+    {
+        this._writer.WriteLine("//! This file is automatically generated by RenderiteGenerator. Please do not modify this file.");
+        this._writer.WriteLine($"// Generated for {this._engineVersion}\n");
+        
+        this._writer.WriteLine("const buffer = @import(\"buffer.zig\");");
+        this._writer.WriteLine("const SharedMemoryBufferDescriptor = buffer.SharedMemoryBufferDescriptor;");
+        this._writer.WriteLine();
+        this._writer.WriteLine("const serialization = @import(\"serialization.zig\");");
+        this._writer.WriteLine("const IpcDeserializer = serialization.IpcDeserializer;");
+        this._writer.WriteLine("const IpcSerializer = serialization.IpcSerializer;");
+        this._writer.WriteLine();
+        this._writer.WriteLine("const math = @import(\"math\");");
+        this._writer.WriteLine();
+        
+        Console.WriteLine("Generating polymorphic entity enums...");
+        this._writer.WriteLine("// Polymorphic entity enums (PolymorphicMemoryPackableEntity<T>)");
+        Type polymorphicType = this._types.First(t => t.Name == "PolymorphicMemoryPackableEntity`1").GetGenericTypeDefinition();
+        foreach (Type type in this._types.Where(t =>
+                 {
+                     if (t.BaseType is not { IsGenericType: true })
+                         return false;
+                     
+                     return t.BaseType.GetGenericTypeDefinition() == polymorphicType;
+                 }))
+        {
+            this.WritePolymorphicEnum(type);
+        }
+
+        Console.WriteLine("Generating enums...");
+        this._writer.WriteLine("// Generated Enums");
+        foreach (Type type in this._types.Where(t => t.IsEnum))
+        {
+            this._generatedTypes.Add(type);
+            this.WriteEnum(type);
+        }
+
+        Console.WriteLine("Generating structs...");
+        this._writer.WriteLine("// Generated Structs");
+        foreach (Type type in this._types.Where(t => t != this._iMemoryPackable && !t.IsAbstract && t.IsAssignableTo(this._iMemoryPackable)))
+        {
+            this._generatedTypes.Add(type);
+            this.WriteStruct(type, true);
+        }
+        
+        Console.WriteLine("Generating remaining structs...");
+        while (this._typesToGenerate.TryDequeue(out Type? type))
+        {
+            if(this._generatedTypes.Contains(type) || type.IsAbstract)
+                continue;
+            
+            this._generatedTypes.Add(type);
+            this.WriteStruct(type);
+        }
+    }
+
+    private void WriteEnum(Type t)
+    {
+        if(this._verbose)
+            Console.WriteLine($"\tWriting enum {t.FullName}");
+            
+        FieldInfo valueField = t.GetField("value__")!;
+        Type underlyingType = valueField.FieldType;
+
+        Array values = Enum.GetValues(t);
+
+        EnumInfo info = new()
+        {
+            Name = t.Name,
+            UnderlyingType = underlyingType,
+            BitSize = Marshal.SizeOf(underlyingType) * 8,
+        };
+        
+        if (t.DeclaringType != null)
+        {
+            info.Name = t.DeclaringType.Name + '_' + info.Name;
+        }
+
+        List<EnumItemInfo> itemInfo = [];
+        foreach (object value in values)
+        {
+            string? name = value.ToString();
+            Debug.Assert(name != null);
+
+            if(itemInfo.Any(i => i.Name == name))
+                continue;
+
+            // workaround for enums that aren't int. this effectively casts to the underlying enum type
+            object? num = valueField.GetValue(value);
+            Debug.Assert(num != null);
+
+            itemInfo.Add(new EnumItemInfo
+            {
+                Name = name,
+                Value = num,
+            });
+        }
+
+        info.Items = itemInfo;
+        
+        bool flags = t.GetCustomAttribute<FlagsAttribute>() != null;
+
+        if(flags)
+            WriteFlagEnum(t, info);
+        else
+            WriteValueEnum(t, info);
+    }
+
+    private void WriteValueEnum(Type t, EnumInfo info)
+    {
+        this._writer.WriteLine($"pub const {info.Name} = enum({MapToZigType(info.UnderlyingType)}) {{");
+
+        foreach (EnumItemInfo value in info.Items)
+        {
+            this._writer.WriteLine($"\t{value.Name} = {value.Value},");
+        }
+        
+        this._writer.WriteLine("};\n");
+    }
+    
+    private void WriteFlagEnum(Type t, EnumInfo info)
+    {
+        this._writer.WriteLine($"pub const {info.Name} = packed struct({MapToZigType(info.UnderlyingType)}) {{");
+        int bits = info.BitSize;
+        foreach (EnumItemInfo value in info.Items)
+        {
+            if (Convert.ToInt32(value.Value) == 0)
+            {
+                this._writer.WriteLine($"\t// Skipped {value.Name}");
+                continue;
+            }
+            
+            this._writer.WriteLine($"\t{value.Name}: bool, // = {value.Value}");
+            bits--;
+        }
+
+        if (bits != 0)
+        {
+            this._writer.WriteLine($"\tpadding: u{bits} = 0,");
+        }
+
+        this._writer.WriteLine("};\n");
+    }
+    
+    [SuppressMessage("ReSharper", "ConvertIfStatementToConditionalTernaryExpression")]
+    private void WriteStruct(Type type, bool isMemoryPackable = false)
+    {
+        if(this._verbose)
+            Console.WriteLine($"\tWriting struct {type.FullName}");
+        
+        // if we were called from a place that didn't already check if we're a packable, check
+        if (!isMemoryPackable)
+            isMemoryPackable = type.IsAssignableTo(this._iMemoryPackable);
+
+        bool isExplicitPackable = false;
+        if (!isMemoryPackable)
+            isExplicitPackable = (type.Attributes & TypeAttributes.ExplicitLayout) != 0;
+        
+        string structName = type.Name;
+
+        this._writer.WriteLine(isExplicitPackable
+            ? $"pub const {structName} = extern struct {{"
+            : $"pub const {structName} = struct {{");
+
+        FieldInfo[] fields = type.GetFields();
+        foreach (FieldInfo field in fields)
+        {
+            this._writer.WriteLine($"\t{field.Name}: {MapToZigType(field.FieldType)},");
+        }
+
+        if ((isMemoryPackable || isExplicitPackable) && fields.Length > 0)
+        {
+            // ReSharper disable once RedundantAssignment
+            bool generated = false;
+            
+            this._writer.WriteLine();
+            this._writer.WriteLine($"\tpub fn write(self: {structName}, ipc: IpcSerializer) !void {{");
+            if(isExplicitPackable)
+                generated = WriteExplicitPackFunction(type, fields, false);
+            else
+                generated = WritePackFunction(type, type.GetMethod("Pack")!);
+            if(!generated) WritePackDiscard();
+            this._writer.WriteLine("\t}\n");
+            
+            this._writer.WriteLine($"\tpub fn read(ipc: IpcDeserializer) !{structName} {{");
+            this._writer.WriteLine($"\t\tvar self: {structName} = undefined;");
+            if(isExplicitPackable)
+                generated = WriteExplicitPackFunction(type, fields, true);
+            else
+                generated = WritePackFunction(type, type.GetMethod("Unpack")!);
+            if(!generated) WritePackDiscard(false);
+            this._writer.WriteLine("\t\treturn self;");
+            this._writer.WriteLine("\t}");
+        }
+        
+        this._writer.WriteLine("};\n");
+    }
+
+    private bool WritePackFunction(Type type, MethodInfo method)
+    {
+        bool written = false;
+        
+        if(_ilVerbose)
+            this._writer.WriteLine($"\t\t// {method.Name} {type.Name}");
+
+        TypeDefinition typeDef = this._cecilAssembly.MainModule.GetType(type.Namespace + '.' + type.Name);
+        MethodDefinition? methodDef = typeDef.GetMethods().FirstOrDefault(m => m.Name == method.Name);
+        if (methodDef == null)
+        {
+            if (type.BaseType == null)
+            {
+                // this should never happen
+                this._writer.WriteLine($"\t\t// BUG: {type.Name} has no explicitly defined method named {method.Name}, and has no base type");
+                return written;
+            }
+            
+            if(this._ilVerbose)
+                this._writer.WriteLine($"\t\t// NOTE: {type.Name} has no explicitly defined method named {method.Name}, trying {type.BaseType?.Name}.{method.Name}");
+
+            written |= WritePackFunction(type.BaseType!, type.BaseType!.GetMethod(method.Name)!);
+            return written;
+        }
+
+        Queue<string> names = new();
+        foreach (Instruction instruction in methodDef.Body.Instructions)
+        {
+            // if (this._ilVerbose)
+                // this._writer.WriteLine($"\t\t// {instruction} ({instruction.OpCode.FlowControl})");
+
+            FlowControl flow = instruction.OpCode.FlowControl;
+            if (flow != FlowControl.Next && flow != FlowControl.Call && flow != FlowControl.Return)
+            {
+                this._writer.WriteLine($"\t\t// FIXME: Unknown {flow} instruction:\n\t\t// {instruction}");
+            }
+
+            if (instruction.OpCode.Code is Code.Ldfld or Code.Ldflda)
+            {
+                string name = ((FieldReference)instruction.Operand).Name;
+                names.Enqueue(name);
+                // if (this._ilVerbose)
+                    // this._writer.WriteLine($"\t\t// {instruction.OpCode.Code} {name}");
+            }
+
+            if (instruction.OpCode.Code is Code.Call && instruction.Operand is MethodReference callRef)
+            {
+                switch (callRef.Name)
+                {
+                    // Write single value or packed boolean
+                    case "Write" when callRef.Parameters.Count == 1:
+                    case "WriteObject":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\ttry ipc.write(@TypeOf(self.{name}), self.{name});");
+                        written = true;
+                        break;
+                    }
+                    case "Write" when callRef.Parameters.All(p => p.ParameterType.Name == "Boolean"):
+                    {
+                        List<string> paramsList = names.Select(n => "self." + n).ToList();
+                        while (paramsList.Count < 8)
+                        {
+                            paramsList.Add("false");
+                        }
+                        this._writer.WriteLine($"\t\ttry ipc.write{paramsList.Count}PackedBools({string.Join(", ", paramsList)});");
+                        names.Clear();
+                        written = true;
+                        break;
+                    }
+                    // Read single value or packed boolean
+                    case "Read" when callRef.Parameters.Count == 1:
+                    case "ReadObject":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\tself.{name} = try ipc.read(@TypeOf(self.{name}));");
+                        written = true;
+                        break;
+                    }
+                    case "Read" when callRef.Parameters.All(p => p.ParameterType.Name == "Boolean&"):
+                    {
+                        List<string> paramsList = names.Select(n => "self." + n).ToList();
+                        // TODO: ceil to nearest multiple of 8
+                        // FE doesn't pack more than 1 byte at a time so its not particularly important
+                        while (paramsList.Count < 8)
+                        {
+                            paramsList.Add("_");
+                        }
+                        this._writer.WriteLine($"\t\t{string.Join(", ", paramsList)} = try ipc.read{paramsList.Count}PackedBools();");
+                        names.Clear();
+                        written = true;
+                        break;
+                    }
+                    // Write list
+                    case "WriteValueList":
+                    case "WriteObjectList":
+                    case "WriteStringList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\ttry ipc.writeList(@TypeOf(self.{name}), self.{name});");
+                        written = true;
+                        break;
+                    }
+                    case "WriteNestedValueList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\ttry ipc.writeNestedList(@TypeOf(self.{name}), self.{name});");
+                        written = true;
+                        break;
+                    }
+                    case "WritePolymorphicList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\ttry ipc.writePolymorphicList(@TypeOf(self.{name}), self.{name});");
+                        written = true;
+                        break;
+                    }
+                    // Read list
+                    case "ReadValueList":
+                    case "ReadObjectList":
+                    case "ReadStringList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\tself.{name} = try ipc.readList(@TypeOf(self.{name}));");
+                        written = true;
+                        break;
+                    }
+                    case "ReadNestedValueList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\tself.{name} = try ipc.readNestedList(@TypeOf(self.{name}));");
+                        written = true;
+                        break;
+                    }
+                    case "ReadPolymorphicList":
+                    {
+                        string name = names.DequeueLast();
+                        this._writer.WriteLine($"\t\tself.{name} = try ipc.readPolymorphicList(@TypeOf(self.{name}));");
+                        written = true;
+                        break;
+                    }
+                    // Packing methods
+                    case "Pack" or "Unpack":
+                    {
+                        MethodInfo? subMethod = type.BaseType?.GetMethod(method.Name);
+                        if (subMethod == null)
+                        {
+                            this._writer.WriteLine($"\t\t// FIXME: Could not find {method.Name} on {type.BaseType}");
+                            written = true;
+                            continue;
+                        }
+                        written |= WritePackFunction(type.BaseType!, subMethod);
+                        break;
+                    }
+                    default:
+                        this._writer.WriteLine($"\t\t// FIXME: Unknown {callRef.GetType().Name} {callRef}");
+                        break;
+                }
+            }
+        }
+
+        if(_ilVerbose)
+            this._writer.WriteLine($"\t\t// {type.Name} {method.Name.ToLower()}ed");
+
+        return written;
+    }
+    
+    private bool WriteExplicitPackFunction(Type type, FieldInfo[] fields, bool read)
+    {
+        const bool ipcStruct = false;
+        
+        int last = 0;
+        bool written = false;
+        foreach (FieldInfo field in fields)
+        {
+            FieldOffsetAttribute? offset = field.GetCustomAttribute<FieldOffsetAttribute>();
+            Debug.Assert(offset != null);
+
+            Type fieldType = field.FieldType;
+            if (fieldType.IsEnum)
+            {
+                FieldInfo valueField = fieldType.GetField("value__")!;
+                fieldType = valueField.FieldType;
+            }
+
+            int size = Marshal.SizeOf(fieldType);
+            int gap = offset.Value - last;
+
+            if (gap != 0) this._writer.WriteLine($"\t\t// NOTE: field with gap/overlap, {field.Name} = offset:{offset.Value}, size:{size}, gap:{gap}");
+
+            string name = field.Name;
+
+            if (!ipcStruct)
+            {
+                if (read)
+                    this._writer.WriteLine($"\t\tself.{name} = try ipc.read(@TypeOf(self.{name}));");
+                else
+                    this._writer.WriteLine($"\t\ttry ipc.write(@TypeOf(self.{name}), self.{name});");
+
+                written = true;
+            }
+            
+            last = offset.Value + size;
+        }
+
+        if (ipcStruct)
+        {
+            if(read)
+                this._writer.WriteLine($"\t\tself = try ipc.readStruct({type.Name});");
+            else
+                this._writer.WriteLine($"\t\ttry ipc.writeStruct({type.Name}, self);");
+
+            written = true;
+        }
+
+        return written;
+    }
+
+    private void WritePackDiscard(bool self = true)
+    {
+        if (self)
+            this._writer.WriteLine("\t\t_ = self;");
+        else
+            this._writer.WriteLine("\t\t_ = &self; // FIXME: Type not generating any members");
+
+        this._writer.WriteLine("\t\t_ = ipc;");
+    }
+    
+    private void WritePolymorphicEnum(Type type)
+    {
+        if(this._verbose)
+            Console.WriteLine($"\tWriting polymorphic enum {type.FullName}");
+
+        string enumName = type.Name;
+        
+        this._writer.WriteLine($"pub const {enumName}Types = enum(i32) {{");
+
+        TypeDefinition typeDef = this._cecilAssembly.MainModule.GetType(type.FullName);
+        MethodDefinition? ctor = typeDef.GetStaticConstructor();
+        if (ctor == null)
+            throw new Exception($"Couldn't find static constructor for {enumName}");
+
+        List<string> typeNames = [];
+        
+        foreach (Instruction instruction in ctor.Body.Instructions)
+        {
+            if (instruction.OpCode.Code != Code.Ldtoken) continue;
+
+            string name = ((TypeDefinition)instruction.Operand).Name;
+            this._writer.WriteLine($"\t{name},");
+            typeNames.Add(name);
+        }
+        
+        this._writer.WriteLine($"}};\npub const {enumName} = union({enumName}Types) {{");
+        foreach (string name in typeNames)
+        {
+            this._writer.WriteLine($"\t{name}: {name},");
+        }
+        this._writer.WriteLine("};\n");
+    }
+
+    private string MapToZigType(Type type, bool inList = false)
+    {
+        if (type == typeof(string))
+            return "[]const u16";
+        
+        if (type == typeof(byte))
+            return "u8";
+        if (type == typeof(sbyte))
+            return "i8";
+        
+        if (type == typeof(short))
+            return "i16";
+        if (type == typeof(ushort))
+            return "u16";
+        
+        if (type == typeof(int))
+            return "i32";
+        if (type == typeof(uint))
+            return "u32";
+        
+        if (type == typeof(long))
+            return "i64";
+        if (type == typeof(ulong))
+            return "u64";
+        
+        if (type == typeof(Int128))
+            return "i128";
+        if (type == typeof(UInt128))
+            return "u128";
+
+        if (type == typeof(float))
+            return "f32";
+        if (type == typeof(double))
+            return "f64";
+
+        if (type == typeof(bool))
+            return "bool";
+
+        switch (type.Name)
+        {
+            case "RenderVector2":
+                return "math.Vector2f";
+            case "RenderVector2i":
+                return "math.Vector2i";
+            case "RenderVector3":
+                return "math.Vector3f";
+            case "RenderVector3i":
+                return "math.Vector3i";
+            case "RenderVector4":
+                return "math.Vector4f";
+            case "RenderVector4i":
+                return "math.Vector4i";
+            case "RenderQuaternion":
+                return "math.Quaternionf";
+            case "RenderMatrix4x4":
+                return "math.Matrix4x4f";
+        }
+
+        // set inList so we don't end up with nullable types in lists
+        if (typeof(IEnumerable).IsAssignableFrom(type))
+            return $"[]const {MapToZigType(type.GenericTypeArguments.First(), true)}";
+        
+        if(type.Name == "Nullable`1")
+            return $"?{MapToZigType(type.GenericTypeArguments.First(), inList)}";
+
+        // all c# classes are nullable
+        // they're also always written with WriteObject as Write only works for unmanaged types
+        if (type.IsClass && !inList)
+            return $"?{type.Name}";
+
+        if (type.Name.StartsWith("SharedMemoryBufferDescriptor"))
+        {
+            _ = MapToZigType(type.GenericTypeArguments.First(), inList);
+            return "SharedMemoryBufferDescriptor";
+        }
+
+        if (type.IsGenericType)
+            return $"{type.Name.Remove(type.Name.IndexOf('`'))}({string.Join(", ", type.GenericTypeArguments.Select(t => MapToZigType(t, inList)))})";
+        
+        QueueType(type);
+
+        return type.Name;
+    }
+
+    private void QueueType(Type type)
+    {
+        if (this._generatedTypes.Contains(type)) return;
+
+        if (type.Assembly == this._assembly)
+            this._typesToGenerate.Enqueue(type);
+    }
+
+    public void Dispose()
+    {
+        this._writer.Flush();
+        this._fileStream.Flush();
+        
+        this._writer.Dispose();
+        this._fileStream.Dispose();
+        GC.SuppressFinalize(this);
+    }
+}
