+++ renderite/generator/RenderiteGenerator/Logging/LogCategory.cs
@@ -2,6 +2,7 @@
 
 public enum LogCategory
 {
+    Startup,
     Settings,
     Generator,
     Parsing,
