+++ renderite/shared.zig
@@ -1,4 +1,4 @@
-//! This file is automatically generated by RenderiteGenerator. Please do not modify this file.
+// This file is automatically generated by RenderiteGenerator. Please do not modify this file.
 // Generated for FrooxEngine, Version=2025.8.11.521, Culture=neutral, PublicKeyToken=null
 
 const buffer = @import("buffer.zig");
@@ -10,4896 +10,4 @@ const IpcSerializer = serialization.IpcSerializer;
 
 const math = @import("math");
 
-// Polymorphic entity enums (PolymorphicMemoryPackableEntity<T>)
-pub const VR_ControllerStateTypes = enum(i32) {
-    CosmosControllerState,
-    GenericControllerState,
-    HP_ReverbControllerState,
-    IndexControllerState,
-    PicoNeo2ControllerState,
-    TouchControllerState,
-    ViveControllerState,
-    WindowsMR_ControllerState,
-};
-pub const VR_ControllerState = union(VR_ControllerStateTypes) {
-    CosmosControllerState: CosmosControllerState,
-    GenericControllerState: GenericControllerState,
-    HP_ReverbControllerState: HP_ReverbControllerState,
-    IndexControllerState: IndexControllerState,
-    PicoNeo2ControllerState: PicoNeo2ControllerState,
-    TouchControllerState: TouchControllerState,
-    ViveControllerState: ViveControllerState,
-    WindowsMR_ControllerState: WindowsMR_ControllerState,
-};
-
-pub const RendererCommandTypes = enum(i32) {
-    RendererInitData,
-    RendererInitResult,
-    RendererInitProgressUpdate,
-    RendererInitFinalizeData,
-    RendererShutdownRequest,
-    RendererShutdown,
-    KeepAlive,
-    RendererParentWindow,
-    FrameStartData,
-    FrameSubmitData,
-    PostProcessingConfig,
-    QualityConfig,
-    ResolutionConfig,
-    DesktopConfig,
-    GaussianSplatConfig,
-    MeshUploadData,
-    MeshUnload,
-    MeshUploadResult,
-    ShaderUpload,
-    ShaderUnload,
-    ShaderUploadResult,
-    MaterialPropertyIdRequest,
-    MaterialPropertyIdResult,
-    MaterialsUpdateBatch,
-    MaterialsUpdateBatchResult,
-    SetTexture2DFormat,
-    SetTexture2DProperties,
-    SetTexture2DData,
-    SetTexture2DResult,
-    UnloadTexture2D,
-    SetTexture3DFormat,
-    SetTexture3DProperties,
-    SetTexture3DData,
-    SetTexture3DResult,
-    UnloadTexture3D,
-    SetCubemapFormat,
-    SetCubemapProperties,
-    SetCubemapData,
-    SetCubemapResult,
-    UnloadCubemap,
-    SetRenderTextureFormat,
-    RenderTextureResult,
-    UnloadRenderTexture,
-    SetDesktopTextureProperties,
-    DesktopTexturePropertiesUpdate,
-    UnloadDesktopTexture,
-    PointRenderBufferUpload,
-    PointRenderBufferConsumed,
-    PointRenderBufferUnload,
-    TrailRenderBufferUpload,
-    TrailRenderBufferConsumed,
-    TrailRenderBufferUnload,
-    GaussianSplatUploadRaw,
-    GaussianSplatUploadEncoded,
-    GaussianSplatResult,
-    UnloadGaussianSplat,
-    LightsBufferRendererSubmission,
-    LightsBufferRendererConsumed,
-    ReflectionProbeRenderResult,
-    VideoTextureLoad,
-    VideoTextureUpdate,
-    VideoTextureReady,
-    VideoTextureChanged,
-    VideoTextureProperties,
-    VideoTextureStartAudioTrack,
-    UnloadVideoTexture,
-};
-pub const RendererCommand = union(RendererCommandTypes) {
-    RendererInitData: RendererInitData,
-    RendererInitResult: RendererInitResult,
-    RendererInitProgressUpdate: RendererInitProgressUpdate,
-    RendererInitFinalizeData: RendererInitFinalizeData,
-    RendererShutdownRequest: RendererShutdownRequest,
-    RendererShutdown: RendererShutdown,
-    KeepAlive: KeepAlive,
-    RendererParentWindow: RendererParentWindow,
-    FrameStartData: FrameStartData,
-    FrameSubmitData: FrameSubmitData,
-    PostProcessingConfig: PostProcessingConfig,
-    QualityConfig: QualityConfig,
-    ResolutionConfig: ResolutionConfig,
-    DesktopConfig: DesktopConfig,
-    GaussianSplatConfig: GaussianSplatConfig,
-    MeshUploadData: MeshUploadData,
-    MeshUnload: MeshUnload,
-    MeshUploadResult: MeshUploadResult,
-    ShaderUpload: ShaderUpload,
-    ShaderUnload: ShaderUnload,
-    ShaderUploadResult: ShaderUploadResult,
-    MaterialPropertyIdRequest: MaterialPropertyIdRequest,
-    MaterialPropertyIdResult: MaterialPropertyIdResult,
-    MaterialsUpdateBatch: MaterialsUpdateBatch,
-    MaterialsUpdateBatchResult: MaterialsUpdateBatchResult,
-    SetTexture2DFormat: SetTexture2DFormat,
-    SetTexture2DProperties: SetTexture2DProperties,
-    SetTexture2DData: SetTexture2DData,
-    SetTexture2DResult: SetTexture2DResult,
-    UnloadTexture2D: UnloadTexture2D,
-    SetTexture3DFormat: SetTexture3DFormat,
-    SetTexture3DProperties: SetTexture3DProperties,
-    SetTexture3DData: SetTexture3DData,
-    SetTexture3DResult: SetTexture3DResult,
-    UnloadTexture3D: UnloadTexture3D,
-    SetCubemapFormat: SetCubemapFormat,
-    SetCubemapProperties: SetCubemapProperties,
-    SetCubemapData: SetCubemapData,
-    SetCubemapResult: SetCubemapResult,
-    UnloadCubemap: UnloadCubemap,
-    SetRenderTextureFormat: SetRenderTextureFormat,
-    RenderTextureResult: RenderTextureResult,
-    UnloadRenderTexture: UnloadRenderTexture,
-    SetDesktopTextureProperties: SetDesktopTextureProperties,
-    DesktopTexturePropertiesUpdate: DesktopTexturePropertiesUpdate,
-    UnloadDesktopTexture: UnloadDesktopTexture,
-    PointRenderBufferUpload: PointRenderBufferUpload,
-    PointRenderBufferConsumed: PointRenderBufferConsumed,
-    PointRenderBufferUnload: PointRenderBufferUnload,
-    TrailRenderBufferUpload: TrailRenderBufferUpload,
-    TrailRenderBufferConsumed: TrailRenderBufferConsumed,
-    TrailRenderBufferUnload: TrailRenderBufferUnload,
-    GaussianSplatUploadRaw: GaussianSplatUploadRaw,
-    GaussianSplatUploadEncoded: GaussianSplatUploadEncoded,
-    GaussianSplatResult: GaussianSplatResult,
-    UnloadGaussianSplat: UnloadGaussianSplat,
-    LightsBufferRendererSubmission: LightsBufferRendererSubmission,
-    LightsBufferRendererConsumed: LightsBufferRendererConsumed,
-    ReflectionProbeRenderResult: ReflectionProbeRenderResult,
-    VideoTextureLoad: VideoTextureLoad,
-    VideoTextureUpdate: VideoTextureUpdate,
-    VideoTextureReady: VideoTextureReady,
-    VideoTextureChanged: VideoTextureChanged,
-    VideoTextureProperties: VideoTextureProperties,
-    VideoTextureStartAudioTrack: VideoTextureStartAudioTrack,
-    UnloadVideoTexture: UnloadVideoTexture,
-};
-
-// Generated Enums
-pub const ColorProfile = enum(i32) {
-    Linear = 0,
-    sRGB = 1,
-    sRGBAlpha = 2,
-};
-
-pub const GaussianVectorFormat = enum(i32) {
-    Float32 = 0,
-    Norm16 = 1,
-    Norm11 = 2,
-    Norm6 = 3,
-};
-
-pub const GaussianRotationFormat = enum(i32) {
-    PackedNorm10 = 0,
-};
-
-pub const GaussianColorFormat = enum(i32) {
-    Float32x4 = 0,
-    Float16x4 = 1,
-    Norm8x4 = 2,
-    BC7 = 3,
-};
-
-pub const GaussianSHFormat = enum(i32) {
-    Float16 = 0,
-    Norm11 = 1,
-    Norm6 = 2,
-    Cluster64k = 3,
-    Cluster32k = 4,
-    Cluster16k = 5,
-    Cluster8k = 6,
-    Cluster4k = 7,
-};
-
-pub const BlendshapeDataFlags = packed struct(i32) {
-    // Skipped NONE
-    Positions: bool, // = 1
-    Normals: bool, // = 2
-    Tangets: bool, // = 4
-    padding: u29 = 0,
-};
-
-pub const IndexBufferFormat = enum(i32) {
-    UInt16 = 0,
-    UInt32 = 1,
-};
-
-pub const SubmeshTopology = enum(i32) {
-    Points = 0,
-    Triangles = 1,
-};
-
-pub const VertexAttributeFormat = enum(i16) {
-    Float32 = 0,
-    Half16 = 1,
-    UNorm8 = 2,
-    UNorm16 = 3,
-    SInt8 = 4,
-    SInt16 = 5,
-    SInt32 = 6,
-    UInt8 = 7,
-    UInt16 = 8,
-    UInt32 = 9,
-};
-
-pub const VertexAttributeType = enum(i16) {
-    Position = 0,
-    Normal = 1,
-    Tangent = 2,
-    Color = 3,
-    UV0 = 4,
-    UV1 = 5,
-    UV2 = 6,
-    UV3 = 7,
-    UV4 = 8,
-    UV5 = 9,
-    UV6 = 10,
-    UV7 = 11,
-    BoneWeights = 12,
-    BoneIndicies = 13,
-};
-
-pub const TextureAssetType = enum(i32) {
-    Texture2D = 0,
-    Texture3D = 1,
-    Cubemap = 2,
-    RenderTexture = 3,
-    VideoTexture = 4,
-    Desktop = 5,
-};
-
-pub const TextureFilterMode = enum(i32) {
-    Point = 0,
-    Bilinear = 1,
-    Trilinear = 2,
-    Anisotropic = 3,
-};
-
-pub const TextureFormat = enum(i32) {
-    Unknown = 0,
-    Alpha8 = 1,
-    R8 = 2,
-    RGB24 = 16,
-    ARGB32 = 17,
-    RGBA32 = 18,
-    BGRA32 = 19,
-    RGB565 = 24,
-    BGR565 = 25,
-    RGBAHalf = 32,
-    ARGBHalf = 33,
-    RHalf = 34,
-    RGHalf = 35,
-    RGBAFloat = 48,
-    ARGBFloat = 49,
-    RFloat = 50,
-    RGFloat = 51,
-    BC1 = 64,
-    BC2 = 65,
-    BC3 = 66,
-    BC4 = 67,
-    BC5 = 68,
-    BC6H = 69,
-    BC7 = 70,
-    ETC2_RGB = 96,
-    ETC2_RGBA1 = 97,
-    ETC2_RGBA8 = 98,
-    ASTC_4x4 = 128,
-    ASTC_5x5 = 129,
-    ASTC_6x6 = 130,
-    ASTC_8x8 = 131,
-    ASTC_10x10 = 132,
-    ASTC_12x12 = 133,
-};
-
-pub const TextureType = enum(i32) {
-    Texture2D = 0,
-    Cubemap = 1,
-    Texture3D = 2,
-};
-
-pub const TextureWrapMode = enum(i32) {
-    Repeat = 0,
-    Clamp = 1,
-    Mirror = 2,
-    MirrorOnce = 3,
-};
-
-pub const Key = enum(i32) {
-    None = 0,
-    Backspace = 8,
-    Tab = 9,
-    Clear = 12,
-    Return = 13,
-    Pause = 19,
-    Escape = 27,
-    Space = 32,
-    Exclaim = 33,
-    DoubleQuote = 34,
-    Hash = 35,
-    Dollar = 36,
-    Ampersand = 38,
-    Quote = 39,
-    LeftParenthesis = 40,
-    RightParenthesis = 41,
-    Asterisk = 42,
-    Plus = 43,
-    Comma = 44,
-    Minus = 45,
-    Period = 46,
-    Slash = 47,
-    Alpha0 = 48,
-    Alpha1 = 49,
-    Alpha2 = 50,
-    Alpha3 = 51,
-    Alpha4 = 52,
-    Alpha5 = 53,
-    Alpha6 = 54,
-    Alpha7 = 55,
-    Alpha8 = 56,
-    Alpha9 = 57,
-    Colon = 58,
-    Semicolon = 59,
-    Less = 60,
-    Equals = 61,
-    Greater = 62,
-    Question = 63,
-    At = 64,
-    VerticalBar = 65,
-    LeftBracket = 91,
-    Backslash = 92,
-    RightBracket = 93,
-    Caret = 94,
-    Underscore = 95,
-    BackQuote = 96,
-    A = 97,
-    B = 98,
-    C = 99,
-    D = 100,
-    E = 101,
-    F = 102,
-    G = 103,
-    H = 104,
-    I = 105,
-    J = 106,
-    K = 107,
-    L = 108,
-    M = 109,
-    N = 110,
-    O = 111,
-    P = 112,
-    Q = 113,
-    R = 114,
-    S = 115,
-    T = 116,
-    U = 117,
-    V = 118,
-    W = 119,
-    X = 120,
-    Y = 121,
-    Z = 122,
-    Percent = 123,
-    Tilde = 124,
-    LeftBrace = 125,
-    RightBrace = 126,
-    Delete = 127,
-    Keypad0 = 256,
-    Keypad1 = 257,
-    Keypad2 = 258,
-    Keypad3 = 259,
-    Keypad4 = 260,
-    Keypad5 = 261,
-    Keypad6 = 262,
-    Keypad7 = 263,
-    Keypad8 = 264,
-    Keypad9 = 265,
-    KeypadPeriod = 266,
-    KeypadDivide = 267,
-    KeypadMultiply = 268,
-    KeypadMinus = 269,
-    KeypadPlus = 270,
-    KeypadEnter = 271,
-    KeypadEquals = 272,
-    UpArrow = 273,
-    DownArrow = 274,
-    RightArrow = 275,
-    LeftArrow = 276,
-    Insert = 277,
-    Home = 278,
-    End = 279,
-    PageUp = 280,
-    PageDown = 281,
-    F1 = 282,
-    F2 = 283,
-    F3 = 284,
-    F4 = 285,
-    F5 = 286,
-    F6 = 287,
-    F7 = 288,
-    F8 = 289,
-    F9 = 290,
-    F10 = 291,
-    F11 = 292,
-    F12 = 293,
-    F13 = 294,
-    F14 = 295,
-    F15 = 296,
-    Numlock = 300,
-    CapsLock = 301,
-    ScrollLock = 302,
-    RightShift = 303,
-    LeftShift = 304,
-    RightControl = 305,
-    LeftControl = 306,
-    RightAlt = 307,
-    LeftAlt = 308,
-    RightApple = 309,
-    LeftApple = 310,
-    LeftWindows = 311,
-    RightWindows = 312,
-    AltGr = 313,
-    Help = 315,
-    Print = 316,
-    SysReq = 317,
-    Break = 318,
-    Menu = 319,
-    Shift = 512,
-    Control = 513,
-    Alt = 514,
-    Windows = 515,
-};
-
-pub const RectOrientation = enum(i32) {
-    Default = 0,
-    Clockwise90 = 1,
-    UpsideDown180 = 2,
-    CounterClockwise90 = 3,
-};
-
-pub const MaterialPropertyUpdateType = enum(u8) {
-    SelectTarget = 0,
-    SetShader = 1,
-    SetRenderQueue = 2,
-    SetInstancing = 3,
-    SetRenderType = 4,
-    SetFloat = 5,
-    SetFloat4 = 6,
-    SetFloat4x4 = 7,
-    SetFloatArray = 8,
-    SetFloat4Array = 9,
-    SetTexture = 10,
-    UpdateBatchEnd = 11,
-};
-
-pub const MaterialRenderType = enum(i32) {
-    Opaque = 0,
-    TransparentCutout = 1,
-    Transparent = 2,
-};
-
-pub const TextureUpdateResultType = enum(i32) {
-    FormatSet = 0,
-    PropertiesSet = 1,
-    DataUpload = 2,
-};
-
-pub const AntiAliasingMethod = enum(i32) {
-    Off = 0,
-    FXAA = 1,
-    CTAA = 2,
-    SMAA = 3,
-    TAA = 4,
-};
-
-pub const ShadowCascadeMode = enum(i32) {
-    None = 0,
-    TwoCascades = 1,
-    FourCascades = 2,
-};
-
-pub const ShadowResolutionMode = enum(i32) {
-    Low = 0,
-    Medium = 1,
-    High = 2,
-    Ultra = 3,
-};
-
-pub const SkinWeightMode = enum(i32) {
-    OneBone = 0,
-    TwoBones = 1,
-    FourBones = 2,
-    Unlimited = 3,
-};
-
-pub const HeadOutputDevice = enum(i32) {
-    Autodetect = 0,
-    Headless = 1,
-    Screen = 2,
-    Screen360 = 3,
-    StaticCamera = 4,
-    StaticCamera360 = 5,
-    SteamVR = 6,
-    WindowsMR = 7,
-    Oculus = 8,
-    OculusQuest = 9,
-    UNKNOWN = 10,
-};
-
-pub const TouchControllerModel = enum(i8) {
-    CV1 = 0,
-    QuestAndRiftS = 1,
-};
-
-pub const HeadsetConnection = enum(i32) {
-    Wired = 0,
-    WirelessGeneral = 1,
-    WirelessSteamLink = 2,
-};
-
-pub const CameraClearMode = enum(u8) {
-    Skybox = 0,
-    Color = 1,
-    Depth = 2,
-    Nothing = 3,
-};
-
-pub const CameraProjection = enum(u8) {
-    Perspective = 0,
-    Orthographic = 1,
-    Panoramic = 2,
-};
-
-pub const ComputeResult = enum(i32) {
-    Scheduled = 0,
-    Computed = 1,
-    Postpone = 2,
-    Failed = 3,
-};
-
-pub const LayerType = enum(u8) {
-    Hidden = 0,
-    Overlay = 1,
-};
-
-pub const LightType = enum(u8) {
-    Point = 0,
-    Directional = 1,
-    Spot = 2,
-};
-
-pub const ShadowType = enum(u8) {
-    None = 0,
-    Hard = 1,
-    Soft = 2,
-};
-
-pub const ShadowCastMode = enum(u8) {
-    Off = 0,
-    On = 1,
-    ShadowOnly = 2,
-    DoubleSided = 3,
-};
-
-pub const MotionVectorMode = enum(u8) {
-    Camera = 0,
-    Object = 1,
-    NoMotion = 2,
-};
-
-pub const MeshRendererType = enum(u8) {
-    MeshRenderer = 0,
-    SkinnedMeshRenderer = 1,
-};
-
-pub const ReflectionProbeClear = enum(u8) {
-    Skybox = 0,
-    Color = 1,
-};
-
-pub const ReflectionProbeType = enum(u8) {
-    Baked = 0,
-    OnChanges = 1,
-    Realtime = 2,
-};
-
-pub const ReflectionProbeTimeSlicingMode = enum(u8) {
-    AllFacesAtOnce = 0,
-    IndividualFaces = 1,
-    NoTimeSlicing = 2,
-};
-
-pub const BillboardAlignment = enum(u8) {
-    View = 0,
-    Facing = 1,
-    Local = 2,
-    Global = 3,
-    Direction = 4,
-};
-
-pub const MeshAlignment = enum(u8) {
-    View = 0,
-    Facing = 1,
-    Local = 2,
-    Global = 3,
-};
-
-pub const TrailTextureMode = enum(u8) {
-    Stretch = 0,
-    Tile = 1,
-    DistributePerSegment = 2,
-    RepeatPerSegment = 3,
-};
-
-pub const RenderingContext = enum(u8) {
-    UserView = 0,
-    ExternalView = 1,
-    Camera = 2,
-    Mirror = 3,
-    Portal = 4,
-    RenderToAsset = 5,
-};
-
-pub const Chirality = enum(i8) {
-    Left = 0,
-    Right = 1,
-};
-
-pub const BodyNode = enum(i32) {
-    NONE = 0,
-    Root = 1,
-    View = 2,
-    LeftController = 3,
-    RightController = 4,
-    Hips = 5,
-    Spine = 6,
-    Chest = 7,
-    UpperChest = 8,
-    Neck = 9,
-    Head = 10,
-    Jaw = 11,
-    LeftEye = 12,
-    RightEye = 13,
-    LeftShoulder = 14,
-    LeftUpperArm = 15,
-    LeftLowerArm = 16,
-    LeftHand = 17,
-    LeftPalm = 18,
-    LeftThumb_Metacarpal = 19,
-    LeftThumb_Proximal = 20,
-    LeftThumb_Distal = 21,
-    LeftThumb_Tip = 22,
-    LeftIndexFinger_Metacarpal = 23,
-    LeftIndexFinger_Proximal = 24,
-    LeftIndexFinger_Intermediate = 25,
-    LeftIndexFinger_Distal = 26,
-    LeftIndexFinger_Tip = 27,
-    LeftMiddleFinger_Metacarpal = 28,
-    LeftMiddleFinger_Proximal = 29,
-    LeftMiddleFinger_Intermediate = 30,
-    LeftMiddleFinger_Distal = 31,
-    LeftMiddleFinger_Tip = 32,
-    LeftRingFinger_Metacarpal = 33,
-    LeftRingFinger_Proximal = 34,
-    LeftRingFinger_Intermediate = 35,
-    LeftRingFinger_Distal = 36,
-    LeftRingFinger_Tip = 37,
-    LeftPinky_Metacarpal = 38,
-    LeftPinky_Proximal = 39,
-    LeftPinky_Intermediate = 40,
-    LeftPinky_Distal = 41,
-    LeftPinky_Tip = 42,
-    RightShoulder = 43,
-    RightUpperArm = 44,
-    RightLowerArm = 45,
-    RightHand = 46,
-    RightPalm = 47,
-    RightThumb_Metacarpal = 48,
-    RightThumb_Proximal = 49,
-    RightThumb_Distal = 50,
-    RightThumb_Tip = 51,
-    RightIndexFinger_Metacarpal = 52,
-    RightIndexFinger_Proximal = 53,
-    RightIndexFinger_Intermediate = 54,
-    RightIndexFinger_Distal = 55,
-    RightIndexFinger_Tip = 56,
-    RightMiddleFinger_Metacarpal = 57,
-    RightMiddleFinger_Proximal = 58,
-    RightMiddleFinger_Intermediate = 59,
-    RightMiddleFinger_Distal = 60,
-    RightMiddleFinger_Tip = 61,
-    RightRingFinger_Metacarpal = 62,
-    RightRingFinger_Proximal = 63,
-    RightRingFinger_Intermediate = 64,
-    RightRingFinger_Distal = 65,
-    RightRingFinger_Tip = 66,
-    RightPinky_Metacarpal = 67,
-    RightPinky_Proximal = 68,
-    RightPinky_Intermediate = 69,
-    RightPinky_Distal = 70,
-    RIGHT_FINGER_END = 71,
-    LeftUpperLeg = 72,
-    LeftLowerLeg = 73,
-    LeftFoot = 74,
-    LeftToes = 75,
-    RightUpperLeg = 76,
-    RightLowerLeg = 77,
-    RightFoot = 78,
-    RightToes = 79,
-    END = 80,
-};
-
-pub const FingerSegmentType = enum(i32) {
-    Metacarpal = 0,
-    Proximal = 1,
-    Intermediate = 2,
-    Distal = 3,
-    Tip = 4,
-};
-
-pub const FingerType = enum(i32) {
-    Thumb = 0,
-    Index = 1,
-    Middle = 2,
-    Ring = 3,
-    Pinky = 4,
-};
-
-pub const MeshUploadHint_Flag = packed struct(i32) {
-    VertexLayout: bool, // = 1
-    SubmeshLayout: bool, // = 2
-    Geometry: bool, // = 4
-    Positions: bool, // = 8
-    Normals: bool, // = 16
-    Tangents: bool, // = 32
-    Colors: bool, // = 64
-    UV0s: bool, // = 128
-    UV1s: bool, // = 256
-    UV2s: bool, // = 512
-    UV3s: bool, // = 1024
-    UV4s: bool, // = 2048
-    UV5s: bool, // = 4096
-    UV6s: bool, // = 8192
-    UV7s: bool, // = 16384
-    BindPoses: bool, // = 32768
-    BoneWeights: bool, // = 65536
-    Blendshapes: bool, // = 131072
-    Dynamic: bool, // = 262144
-    Readable: bool, // = 524288
-    Debug: bool, // = 1048576
-    padding: u11 = 0,
-};
-
-// Generated Structs
-pub const GaussianSplatResult = struct {
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: GaussianSplatResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GaussianSplatResult {
-        var self: GaussianSplatResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const GaussianSplatUploadEncoded = struct {
-    positionsFormat: GaussianVectorFormat,
-    rotationsFormat: GaussianRotationFormat,
-    scalesFormat: GaussianVectorFormat,
-    colorsFormat: GaussianColorFormat,
-    shFormat: GaussianSHFormat,
-    texture2DtextureAssetId: i32,
-    shIndexesOffset: i32,
-    chunkCount: i32,
-    shBuffer: SharedMemoryBufferDescriptor,
-    chunksBuffer: SharedMemoryBufferDescriptor,
-    splatCount: i32,
-    bounds: RenderBoundingBox,
-    positionsBuffer: SharedMemoryBufferDescriptor,
-    rotationsBuffer: SharedMemoryBufferDescriptor,
-    scalesBuffer: SharedMemoryBufferDescriptor,
-    colorsBuffer: SharedMemoryBufferDescriptor,
-    assetId: i32,
-
-    pub fn write(self: GaussianSplatUploadEncoded, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.splatCount), self.splatCount);
-        try ipc.write(@TypeOf(self.bounds), self.bounds);
-        try ipc.write(@TypeOf(self.positionsBuffer), self.positionsBuffer);
-        try ipc.write(@TypeOf(self.rotationsBuffer), self.rotationsBuffer);
-        try ipc.write(@TypeOf(self.scalesBuffer), self.scalesBuffer);
-        try ipc.write(@TypeOf(self.colorsBuffer), self.colorsBuffer);
-        try ipc.write(@TypeOf(self.positionsFormat), self.positionsFormat);
-        try ipc.write(@TypeOf(self.rotationsFormat), self.rotationsFormat);
-        try ipc.write(@TypeOf(self.scalesFormat), self.scalesFormat);
-        try ipc.write(@TypeOf(self.colorsFormat), self.colorsFormat);
-        try ipc.write(@TypeOf(self.shFormat), self.shFormat);
-        try ipc.write(@TypeOf(self.texture2DtextureAssetId), self.texture2DtextureAssetId);
-        try ipc.write(@TypeOf(self.shIndexesOffset), self.shIndexesOffset);
-        try ipc.write(@TypeOf(self.chunkCount), self.chunkCount);
-        try ipc.write(@TypeOf(self.shBuffer), self.shBuffer);
-        try ipc.write(@TypeOf(self.chunksBuffer), self.chunksBuffer);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GaussianSplatUploadEncoded {
-        var self: GaussianSplatUploadEncoded = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.splatCount = try ipc.read(@TypeOf(self.splatCount));
-        self.bounds = try ipc.read(@TypeOf(self.bounds));
-        self.positionsBuffer = try ipc.read(@TypeOf(self.positionsBuffer));
-        self.rotationsBuffer = try ipc.read(@TypeOf(self.rotationsBuffer));
-        self.scalesBuffer = try ipc.read(@TypeOf(self.scalesBuffer));
-        self.colorsBuffer = try ipc.read(@TypeOf(self.colorsBuffer));
-        self.positionsFormat = try ipc.read(@TypeOf(self.positionsFormat));
-        self.rotationsFormat = try ipc.read(@TypeOf(self.rotationsFormat));
-        self.scalesFormat = try ipc.read(@TypeOf(self.scalesFormat));
-        self.colorsFormat = try ipc.read(@TypeOf(self.colorsFormat));
-        self.shFormat = try ipc.read(@TypeOf(self.shFormat));
-        self.texture2DtextureAssetId = try ipc.read(@TypeOf(self.texture2DtextureAssetId));
-        self.shIndexesOffset = try ipc.read(@TypeOf(self.shIndexesOffset));
-        self.chunkCount = try ipc.read(@TypeOf(self.chunkCount));
-        self.shBuffer = try ipc.read(@TypeOf(self.shBuffer));
-        self.chunksBuffer = try ipc.read(@TypeOf(self.chunksBuffer));
-        return self;
-    }
-};
-
-pub const GaussianSplatUploadRaw = struct {
-    alphasBuffer: SharedMemoryBufferDescriptor,
-    splatCount: i32,
-    bounds: RenderBoundingBox,
-    positionsBuffer: SharedMemoryBufferDescriptor,
-    rotationsBuffer: SharedMemoryBufferDescriptor,
-    scalesBuffer: SharedMemoryBufferDescriptor,
-    colorsBuffer: SharedMemoryBufferDescriptor,
-    assetId: i32,
-
-    pub fn write(self: GaussianSplatUploadRaw, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.splatCount), self.splatCount);
-        try ipc.write(@TypeOf(self.bounds), self.bounds);
-        try ipc.write(@TypeOf(self.positionsBuffer), self.positionsBuffer);
-        try ipc.write(@TypeOf(self.rotationsBuffer), self.rotationsBuffer);
-        try ipc.write(@TypeOf(self.scalesBuffer), self.scalesBuffer);
-        try ipc.write(@TypeOf(self.colorsBuffer), self.colorsBuffer);
-        try ipc.write(@TypeOf(self.alphasBuffer), self.alphasBuffer);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GaussianSplatUploadRaw {
-        var self: GaussianSplatUploadRaw = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.splatCount = try ipc.read(@TypeOf(self.splatCount));
-        self.bounds = try ipc.read(@TypeOf(self.bounds));
-        self.positionsBuffer = try ipc.read(@TypeOf(self.positionsBuffer));
-        self.rotationsBuffer = try ipc.read(@TypeOf(self.rotationsBuffer));
-        self.scalesBuffer = try ipc.read(@TypeOf(self.scalesBuffer));
-        self.colorsBuffer = try ipc.read(@TypeOf(self.colorsBuffer));
-        self.alphasBuffer = try ipc.read(@TypeOf(self.alphasBuffer));
-        return self;
-    }
-};
-
-pub const UnloadGaussianSplat = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadGaussianSplat, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadGaussianSplat {
-        var self: UnloadGaussianSplat = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const MaterialPropertyIdRequest = struct {
-    requestId: i32,
-    propertyNames: []const []const u16,
-
-    pub fn write(self: MaterialPropertyIdRequest, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.requestId), self.requestId);
-        try ipc.writeList(@TypeOf(self.propertyNames), self.propertyNames);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MaterialPropertyIdRequest {
-        var self: MaterialPropertyIdRequest = undefined;
-        self.requestId = try ipc.read(@TypeOf(self.requestId));
-        self.propertyNames = try ipc.readList(@TypeOf(self.propertyNames));
-        return self;
-    }
-};
-
-pub const MaterialPropertyIdResult = struct {
-    requestId: i32,
-    propertyIDs: []const i32,
-
-    pub fn write(self: MaterialPropertyIdResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.requestId), self.requestId);
-        try ipc.writeList(@TypeOf(self.propertyIDs), self.propertyIDs);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MaterialPropertyIdResult {
-        var self: MaterialPropertyIdResult = undefined;
-        self.requestId = try ipc.read(@TypeOf(self.requestId));
-        self.propertyIDs = try ipc.readList(@TypeOf(self.propertyIDs));
-        return self;
-    }
-};
-
-pub const MaterialsUpdateBatch = struct {
-    updateBatchId: i32,
-    materialRemovals: SharedMemoryBufferDescriptor,
-    materialPropertyBlockRemovals: SharedMemoryBufferDescriptor,
-    materialUpdates: []const SharedMemoryBufferDescriptor,
-    materialUpdateCount: i32,
-    intBuffers: []const SharedMemoryBufferDescriptor,
-    floatBuffers: []const SharedMemoryBufferDescriptor,
-    float4Buffers: []const SharedMemoryBufferDescriptor,
-    matrixBuffers: []const SharedMemoryBufferDescriptor,
-    instanceChangedBuffer: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: MaterialsUpdateBatch, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.updateBatchId), self.updateBatchId);
-        try ipc.write(@TypeOf(self.materialRemovals), self.materialRemovals);
-        try ipc.write(@TypeOf(self.materialPropertyBlockRemovals), self.materialPropertyBlockRemovals);
-        try ipc.writeList(@TypeOf(self.materialUpdates), self.materialUpdates);
-        try ipc.write(@TypeOf(self.materialUpdateCount), self.materialUpdateCount);
-        try ipc.writeList(@TypeOf(self.intBuffers), self.intBuffers);
-        try ipc.writeList(@TypeOf(self.floatBuffers), self.floatBuffers);
-        try ipc.writeList(@TypeOf(self.float4Buffers), self.float4Buffers);
-        try ipc.writeList(@TypeOf(self.matrixBuffers), self.matrixBuffers);
-        try ipc.write(@TypeOf(self.instanceChangedBuffer), self.instanceChangedBuffer);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MaterialsUpdateBatch {
-        var self: MaterialsUpdateBatch = undefined;
-        self.updateBatchId = try ipc.read(@TypeOf(self.updateBatchId));
-        self.materialRemovals = try ipc.read(@TypeOf(self.materialRemovals));
-        self.materialPropertyBlockRemovals = try ipc.read(@TypeOf(self.materialPropertyBlockRemovals));
-        self.materialUpdates = try ipc.readList(@TypeOf(self.materialUpdates));
-        self.materialUpdateCount = try ipc.read(@TypeOf(self.materialUpdateCount));
-        self.intBuffers = try ipc.readList(@TypeOf(self.intBuffers));
-        self.floatBuffers = try ipc.readList(@TypeOf(self.floatBuffers));
-        self.float4Buffers = try ipc.readList(@TypeOf(self.float4Buffers));
-        self.matrixBuffers = try ipc.readList(@TypeOf(self.matrixBuffers));
-        self.instanceChangedBuffer = try ipc.read(@TypeOf(self.instanceChangedBuffer));
-        return self;
-    }
-};
-
-pub const MaterialsUpdateBatchResult = struct {
-    updateBatchId: i32,
-
-    pub fn write(self: MaterialsUpdateBatchResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.updateBatchId), self.updateBatchId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MaterialsUpdateBatchResult {
-        var self: MaterialsUpdateBatchResult = undefined;
-        self.updateBatchId = try ipc.read(@TypeOf(self.updateBatchId));
-        return self;
-    }
-};
-
-pub const MeshUnload = struct {
-    assetId: i32,
-
-    pub fn write(self: MeshUnload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshUnload {
-        var self: MeshUnload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const MeshUploadData = struct {
-    highPriority: bool,
-    buffer: SharedMemoryBufferDescriptor,
-    vertexCount: i32,
-    boneWeightCount: i32,
-    boneCount: i32,
-    indexBufferFormat: IndexBufferFormat,
-    vertexAttributes: []const VertexAttributeDescriptor,
-    submeshes: []const SubmeshBufferDescriptor,
-    blendshapeBuffers: []const BlendshapeBufferDescriptor,
-    uploadHint: MeshUploadHint,
-    bounds: RenderBoundingBox,
-    assetId: i32,
-
-    pub fn write(self: MeshUploadData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-        try ipc.write(@TypeOf(self.buffer), self.buffer);
-        try ipc.write(@TypeOf(self.vertexCount), self.vertexCount);
-        try ipc.write(@TypeOf(self.boneWeightCount), self.boneWeightCount);
-        try ipc.write(@TypeOf(self.boneCount), self.boneCount);
-        try ipc.write(@TypeOf(self.indexBufferFormat), self.indexBufferFormat);
-        try ipc.writeList(@TypeOf(self.vertexAttributes), self.vertexAttributes);
-        try ipc.writeList(@TypeOf(self.submeshes), self.submeshes);
-        try ipc.writeList(@TypeOf(self.blendshapeBuffers), self.blendshapeBuffers);
-        try ipc.write(@TypeOf(self.uploadHint), self.uploadHint);
-        try ipc.write(@TypeOf(self.bounds), self.bounds);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshUploadData {
-        var self: MeshUploadData = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        self.buffer = try ipc.read(@TypeOf(self.buffer));
-        self.vertexCount = try ipc.read(@TypeOf(self.vertexCount));
-        self.boneWeightCount = try ipc.read(@TypeOf(self.boneWeightCount));
-        self.boneCount = try ipc.read(@TypeOf(self.boneCount));
-        self.indexBufferFormat = try ipc.read(@TypeOf(self.indexBufferFormat));
-        self.vertexAttributes = try ipc.readList(@TypeOf(self.vertexAttributes));
-        self.submeshes = try ipc.readList(@TypeOf(self.submeshes));
-        self.blendshapeBuffers = try ipc.readList(@TypeOf(self.blendshapeBuffers));
-        self.uploadHint = try ipc.read(@TypeOf(self.uploadHint));
-        self.bounds = try ipc.read(@TypeOf(self.bounds));
-        return self;
-    }
-};
-
-pub const MeshUploadResult = struct {
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: MeshUploadResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshUploadResult {
-        var self: MeshUploadResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const PointRenderBufferConsumed = struct {
-    assetId: i32,
-
-    pub fn write(self: PointRenderBufferConsumed, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PointRenderBufferConsumed {
-        var self: PointRenderBufferConsumed = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const PointRenderBufferUnload = struct {
-    assetId: i32,
-
-    pub fn write(self: PointRenderBufferUnload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PointRenderBufferUnload {
-        var self: PointRenderBufferUnload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const PointRenderBufferUpload = struct {
-    count: i32,
-    positionsOffset: i32,
-    rotationsOffset: i32,
-    sizesOffset: i32,
-    colorsOffset: i32,
-    frameIndexesOffset: i32,
-    frameGridSize: math.Vector2i,
-    buffer: SharedMemoryBufferDescriptor,
-    assetId: i32,
-
-    pub fn write(self: PointRenderBufferUpload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.buffer), self.buffer);
-        try ipc.write(@TypeOf(self.count), self.count);
-        try ipc.write(@TypeOf(self.positionsOffset), self.positionsOffset);
-        try ipc.write(@TypeOf(self.rotationsOffset), self.rotationsOffset);
-        try ipc.write(@TypeOf(self.sizesOffset), self.sizesOffset);
-        try ipc.write(@TypeOf(self.colorsOffset), self.colorsOffset);
-        try ipc.write(@TypeOf(self.frameIndexesOffset), self.frameIndexesOffset);
-        try ipc.write(@TypeOf(self.frameGridSize), self.frameGridSize);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PointRenderBufferUpload {
-        var self: PointRenderBufferUpload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.buffer = try ipc.read(@TypeOf(self.buffer));
-        self.count = try ipc.read(@TypeOf(self.count));
-        self.positionsOffset = try ipc.read(@TypeOf(self.positionsOffset));
-        self.rotationsOffset = try ipc.read(@TypeOf(self.rotationsOffset));
-        self.sizesOffset = try ipc.read(@TypeOf(self.sizesOffset));
-        self.colorsOffset = try ipc.read(@TypeOf(self.colorsOffset));
-        self.frameIndexesOffset = try ipc.read(@TypeOf(self.frameIndexesOffset));
-        self.frameGridSize = try ipc.read(@TypeOf(self.frameGridSize));
-        return self;
-    }
-};
-
-pub const TrailRenderBufferConsumed = struct {
-    assetId: i32,
-
-    pub fn write(self: TrailRenderBufferConsumed, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrailRenderBufferConsumed {
-        var self: TrailRenderBufferConsumed = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const TrailRenderBufferUnload = struct {
-    assetId: i32,
-
-    pub fn write(self: TrailRenderBufferUnload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrailRenderBufferUnload {
-        var self: TrailRenderBufferUnload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const TrailRenderBufferUpload = struct {
-    trailsCount: i32,
-    trailPointCount: i32,
-    trailsOffset: i32,
-    positionsOffset: i32,
-    colorsOffset: i32,
-    sizesOffset: i32,
-    buffer: SharedMemoryBufferDescriptor,
-    assetId: i32,
-
-    pub fn write(self: TrailRenderBufferUpload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.buffer), self.buffer);
-        try ipc.write(@TypeOf(self.trailsCount), self.trailsCount);
-        try ipc.write(@TypeOf(self.trailPointCount), self.trailPointCount);
-        try ipc.write(@TypeOf(self.trailsOffset), self.trailsOffset);
-        try ipc.write(@TypeOf(self.positionsOffset), self.positionsOffset);
-        try ipc.write(@TypeOf(self.colorsOffset), self.colorsOffset);
-        try ipc.write(@TypeOf(self.sizesOffset), self.sizesOffset);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrailRenderBufferUpload {
-        var self: TrailRenderBufferUpload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.buffer = try ipc.read(@TypeOf(self.buffer));
-        self.trailsCount = try ipc.read(@TypeOf(self.trailsCount));
-        self.trailPointCount = try ipc.read(@TypeOf(self.trailPointCount));
-        self.trailsOffset = try ipc.read(@TypeOf(self.trailsOffset));
-        self.positionsOffset = try ipc.read(@TypeOf(self.positionsOffset));
-        self.colorsOffset = try ipc.read(@TypeOf(self.colorsOffset));
-        self.sizesOffset = try ipc.read(@TypeOf(self.sizesOffset));
-        return self;
-    }
-};
-
-pub const ShaderUnload = struct {
-    assetId: i32,
-
-    pub fn write(self: ShaderUnload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ShaderUnload {
-        var self: ShaderUnload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const ShaderUpload = struct {
-    file: []const u16,
-    assetId: i32,
-
-    pub fn write(self: ShaderUpload, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.file), self.file);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ShaderUpload {
-        var self: ShaderUpload = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.file = try ipc.read(@TypeOf(self.file));
-        return self;
-    }
-};
-
-pub const ShaderUploadResult = struct {
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: ShaderUploadResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ShaderUploadResult {
-        var self: ShaderUploadResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const SetCubemapData = struct {
-    data: SharedMemoryBufferDescriptor,
-    startMipLevel: i32,
-    mipMapSizes: []const math.Vector2i,
-    mipStarts: []const []const i32,
-    flipY: bool,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetCubemapData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.data), self.data);
-        try ipc.write(@TypeOf(self.startMipLevel), self.startMipLevel);
-        try ipc.writeList(@TypeOf(self.mipMapSizes), self.mipMapSizes);
-        try ipc.writeNestedList(@TypeOf(self.mipStarts), self.mipStarts);
-        try ipc.write(@TypeOf(self.flipY), self.flipY);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetCubemapData {
-        var self: SetCubemapData = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.data = try ipc.read(@TypeOf(self.data));
-        self.startMipLevel = try ipc.read(@TypeOf(self.startMipLevel));
-        self.mipMapSizes = try ipc.readList(@TypeOf(self.mipMapSizes));
-        self.mipStarts = try ipc.readNestedList(@TypeOf(self.mipStarts));
-        self.flipY = try ipc.read(@TypeOf(self.flipY));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        return self;
-    }
-};
-
-pub const SetCubemapFormat = struct {
-    size: i32,
-    mipmapCount: i32,
-    format: TextureFormat,
-    profile: ColorProfile,
-    assetId: i32,
-
-    pub fn write(self: SetCubemapFormat, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.size), self.size);
-        try ipc.write(@TypeOf(self.mipmapCount), self.mipmapCount);
-        try ipc.write(@TypeOf(self.format), self.format);
-        try ipc.write(@TypeOf(self.profile), self.profile);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetCubemapFormat {
-        var self: SetCubemapFormat = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.size = try ipc.read(@TypeOf(self.size));
-        self.mipmapCount = try ipc.read(@TypeOf(self.mipmapCount));
-        self.format = try ipc.read(@TypeOf(self.format));
-        self.profile = try ipc.read(@TypeOf(self.profile));
-        return self;
-    }
-};
-
-pub const SetCubemapProperties = struct {
-    filterMode: TextureFilterMode,
-    anisoLevel: i32,
-    mipmapBias: f32,
-    applyImmediatelly: bool,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetCubemapProperties, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.filterMode), self.filterMode);
-        try ipc.write(@TypeOf(self.anisoLevel), self.anisoLevel);
-        try ipc.write(@TypeOf(self.mipmapBias), self.mipmapBias);
-        try ipc.write(@TypeOf(self.applyImmediatelly), self.applyImmediatelly);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetCubemapProperties {
-        var self: SetCubemapProperties = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.filterMode = try ipc.read(@TypeOf(self.filterMode));
-        self.anisoLevel = try ipc.read(@TypeOf(self.anisoLevel));
-        self.mipmapBias = try ipc.read(@TypeOf(self.mipmapBias));
-        self.applyImmediatelly = try ipc.read(@TypeOf(self.applyImmediatelly));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        return self;
-    }
-};
-
-pub const SetCubemapResult = struct {
-    type: TextureUpdateResultType,
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: SetCubemapResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.type), self.type);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetCubemapResult {
-        var self: SetCubemapResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.type = try ipc.read(@TypeOf(self.type));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const UnloadCubemap = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadCubemap, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadCubemap {
-        var self: UnloadCubemap = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const DesktopTexturePropertiesUpdate = struct {
-    size: math.Vector2i,
-    assetId: i32,
-
-    pub fn write(self: DesktopTexturePropertiesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.size), self.size);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !DesktopTexturePropertiesUpdate {
-        var self: DesktopTexturePropertiesUpdate = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.size = try ipc.read(@TypeOf(self.size));
-        return self;
-    }
-};
-
-pub const SetDesktopTextureProperties = struct {
-    displayIndex: i32,
-    assetId: i32,
-
-    pub fn write(self: SetDesktopTextureProperties, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.displayIndex), self.displayIndex);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetDesktopTextureProperties {
-        var self: SetDesktopTextureProperties = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.displayIndex = try ipc.read(@TypeOf(self.displayIndex));
-        return self;
-    }
-};
-
-pub const UnloadDesktopTexture = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadDesktopTexture, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadDesktopTexture {
-        var self: UnloadDesktopTexture = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const RenderTextureResult = struct {
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: RenderTextureResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderTextureResult {
-        var self: RenderTextureResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const SetRenderTextureFormat = struct {
-    size: math.Vector2i,
-    depth: i32,
-    filterMode: TextureFilterMode,
-    anisoLevel: i32,
-    wrapU: TextureWrapMode,
-    wrapV: TextureWrapMode,
-    assetId: i32,
-
-    pub fn write(self: SetRenderTextureFormat, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.size), self.size);
-        try ipc.write(@TypeOf(self.depth), self.depth);
-        try ipc.write(@TypeOf(self.filterMode), self.filterMode);
-        try ipc.write(@TypeOf(self.anisoLevel), self.anisoLevel);
-        try ipc.write(@TypeOf(self.wrapU), self.wrapU);
-        try ipc.write(@TypeOf(self.wrapV), self.wrapV);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetRenderTextureFormat {
-        var self: SetRenderTextureFormat = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.size = try ipc.read(@TypeOf(self.size));
-        self.depth = try ipc.read(@TypeOf(self.depth));
-        self.filterMode = try ipc.read(@TypeOf(self.filterMode));
-        self.anisoLevel = try ipc.read(@TypeOf(self.anisoLevel));
-        self.wrapU = try ipc.read(@TypeOf(self.wrapU));
-        self.wrapV = try ipc.read(@TypeOf(self.wrapV));
-        return self;
-    }
-};
-
-pub const UnloadRenderTexture = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadRenderTexture, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadRenderTexture {
-        var self: UnloadRenderTexture = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const SetTexture2DData = struct {
-    data: SharedMemoryBufferDescriptor,
-    startMipLevel: i32,
-    mipMapSizes: []const math.Vector2i,
-    mipStarts: []const i32,
-    flipY: bool,
-    hint: TextureUploadHint,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture2DData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.data), self.data);
-        try ipc.write(@TypeOf(self.startMipLevel), self.startMipLevel);
-        try ipc.writeList(@TypeOf(self.mipMapSizes), self.mipMapSizes);
-        try ipc.writeList(@TypeOf(self.mipStarts), self.mipStarts);
-        try ipc.write(@TypeOf(self.flipY), self.flipY);
-        try ipc.write(@TypeOf(self.hint), self.hint);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture2DData {
-        var self: SetTexture2DData = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.data = try ipc.read(@TypeOf(self.data));
-        self.startMipLevel = try ipc.read(@TypeOf(self.startMipLevel));
-        self.mipMapSizes = try ipc.readList(@TypeOf(self.mipMapSizes));
-        self.mipStarts = try ipc.readList(@TypeOf(self.mipStarts));
-        self.flipY = try ipc.read(@TypeOf(self.flipY));
-        self.hint = try ipc.read(@TypeOf(self.hint));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        return self;
-    }
-};
-
-pub const SetTexture2DFormat = struct {
-    width: i32,
-    height: i32,
-    mipmapCount: i32,
-    format: TextureFormat,
-    profile: ColorProfile,
-    assetId: i32,
-
-    pub fn write(self: SetTexture2DFormat, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.width), self.width);
-        try ipc.write(@TypeOf(self.height), self.height);
-        try ipc.write(@TypeOf(self.mipmapCount), self.mipmapCount);
-        try ipc.write(@TypeOf(self.format), self.format);
-        try ipc.write(@TypeOf(self.profile), self.profile);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture2DFormat {
-        var self: SetTexture2DFormat = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.width = try ipc.read(@TypeOf(self.width));
-        self.height = try ipc.read(@TypeOf(self.height));
-        self.mipmapCount = try ipc.read(@TypeOf(self.mipmapCount));
-        self.format = try ipc.read(@TypeOf(self.format));
-        self.profile = try ipc.read(@TypeOf(self.profile));
-        return self;
-    }
-};
-
-pub const SetTexture2DProperties = struct {
-    filterMode: TextureFilterMode,
-    anisoLevel: i32,
-    wrapU: TextureWrapMode,
-    wrapV: TextureWrapMode,
-    mipmapBias: f32,
-    applyImmediatelly: bool,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture2DProperties, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.filterMode), self.filterMode);
-        try ipc.write(@TypeOf(self.anisoLevel), self.anisoLevel);
-        try ipc.write(@TypeOf(self.wrapU), self.wrapU);
-        try ipc.write(@TypeOf(self.wrapV), self.wrapV);
-        try ipc.write(@TypeOf(self.mipmapBias), self.mipmapBias);
-        try ipc.write(@TypeOf(self.applyImmediatelly), self.applyImmediatelly);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture2DProperties {
-        var self: SetTexture2DProperties = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.filterMode = try ipc.read(@TypeOf(self.filterMode));
-        self.anisoLevel = try ipc.read(@TypeOf(self.anisoLevel));
-        self.wrapU = try ipc.read(@TypeOf(self.wrapU));
-        self.wrapV = try ipc.read(@TypeOf(self.wrapV));
-        self.mipmapBias = try ipc.read(@TypeOf(self.mipmapBias));
-        self.applyImmediatelly = try ipc.read(@TypeOf(self.applyImmediatelly));
-        return self;
-    }
-};
-
-pub const SetTexture2DResult = struct {
-    type: TextureUpdateResultType,
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture2DResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.type), self.type);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture2DResult {
-        var self: SetTexture2DResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.type = try ipc.read(@TypeOf(self.type));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const UnloadTexture2D = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadTexture2D, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadTexture2D {
-        var self: UnloadTexture2D = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const SetTexture3DData = struct {
-    data: SharedMemoryBufferDescriptor,
-    hint: Texture3DUploadHint,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture3DData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.data), self.data);
-        try ipc.write(@TypeOf(self.hint), self.hint);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture3DData {
-        var self: SetTexture3DData = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.data = try ipc.read(@TypeOf(self.data));
-        self.hint = try ipc.read(@TypeOf(self.hint));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        return self;
-    }
-};
-
-pub const SetTexture3DFormat = struct {
-    width: i32,
-    height: i32,
-    depth: i32,
-    mipmapCount: i32,
-    format: TextureFormat,
-    profile: ColorProfile,
-    assetId: i32,
-
-    pub fn write(self: SetTexture3DFormat, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.width), self.width);
-        try ipc.write(@TypeOf(self.height), self.height);
-        try ipc.write(@TypeOf(self.depth), self.depth);
-        try ipc.write(@TypeOf(self.mipmapCount), self.mipmapCount);
-        try ipc.write(@TypeOf(self.format), self.format);
-        try ipc.write(@TypeOf(self.profile), self.profile);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture3DFormat {
-        var self: SetTexture3DFormat = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.width = try ipc.read(@TypeOf(self.width));
-        self.height = try ipc.read(@TypeOf(self.height));
-        self.depth = try ipc.read(@TypeOf(self.depth));
-        self.mipmapCount = try ipc.read(@TypeOf(self.mipmapCount));
-        self.format = try ipc.read(@TypeOf(self.format));
-        self.profile = try ipc.read(@TypeOf(self.profile));
-        return self;
-    }
-};
-
-pub const SetTexture3DProperties = struct {
-    filterMode: TextureFilterMode,
-    anisoLevel: i32,
-    wrapU: TextureWrapMode,
-    wrapV: TextureWrapMode,
-    wrapW: TextureWrapMode,
-    applyImmediatelly: bool,
-    highPriority: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture3DProperties, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.filterMode), self.filterMode);
-        try ipc.write(@TypeOf(self.anisoLevel), self.anisoLevel);
-        try ipc.write(@TypeOf(self.wrapU), self.wrapU);
-        try ipc.write(@TypeOf(self.wrapV), self.wrapV);
-        try ipc.write(@TypeOf(self.wrapW), self.wrapW);
-        try ipc.write(@TypeOf(self.applyImmediatelly), self.applyImmediatelly);
-        try ipc.write(@TypeOf(self.highPriority), self.highPriority);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture3DProperties {
-        var self: SetTexture3DProperties = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.filterMode = try ipc.read(@TypeOf(self.filterMode));
-        self.anisoLevel = try ipc.read(@TypeOf(self.anisoLevel));
-        self.wrapU = try ipc.read(@TypeOf(self.wrapU));
-        self.wrapV = try ipc.read(@TypeOf(self.wrapV));
-        self.wrapW = try ipc.read(@TypeOf(self.wrapW));
-        self.applyImmediatelly = try ipc.read(@TypeOf(self.applyImmediatelly));
-        self.highPriority = try ipc.read(@TypeOf(self.highPriority));
-        return self;
-    }
-};
-
-pub const SetTexture3DResult = struct {
-    type: TextureUpdateResultType,
-    instanceChanged: bool,
-    assetId: i32,
-
-    pub fn write(self: SetTexture3DResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.type), self.type);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SetTexture3DResult {
-        var self: SetTexture3DResult = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.type = try ipc.read(@TypeOf(self.type));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        return self;
-    }
-};
-
-pub const UnloadTexture3D = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadTexture3D, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadTexture3D {
-        var self: UnloadTexture3D = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const UnloadVideoTexture = struct {
-    assetId: i32,
-
-    pub fn write(self: UnloadVideoTexture, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !UnloadVideoTexture {
-        var self: UnloadVideoTexture = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const VideoAudioTrack = struct {
-    index: i32,
-    channelCount: i32,
-    sampleRate: i32,
-    name: []const u16,
-    languageCode: []const u16,
-
-    pub fn write(self: VideoAudioTrack, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.index), self.index);
-        try ipc.write(@TypeOf(self.channelCount), self.channelCount);
-        try ipc.write(@TypeOf(self.sampleRate), self.sampleRate);
-        try ipc.write(@TypeOf(self.name), self.name);
-        try ipc.write(@TypeOf(self.languageCode), self.languageCode);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoAudioTrack {
-        var self: VideoAudioTrack = undefined;
-        self.index = try ipc.read(@TypeOf(self.index));
-        self.channelCount = try ipc.read(@TypeOf(self.channelCount));
-        self.sampleRate = try ipc.read(@TypeOf(self.sampleRate));
-        self.name = try ipc.read(@TypeOf(self.name));
-        self.languageCode = try ipc.read(@TypeOf(self.languageCode));
-        return self;
-    }
-};
-
-pub const VideoTextureChanged = struct {
-    assetId: i32,
-
-    pub fn write(self: VideoTextureChanged, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureChanged {
-        var self: VideoTextureChanged = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        return self;
-    }
-};
-
-pub const VideoTextureLoad = struct {
-    source: []const u16,
-    overrideEngine: []const u16,
-    mimeType: []const u16,
-    isStream: bool,
-    audioSystemSampleRate: i32,
-    assetId: i32,
-
-    pub fn write(self: VideoTextureLoad, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.source), self.source);
-        try ipc.write(@TypeOf(self.overrideEngine), self.overrideEngine);
-        try ipc.write(@TypeOf(self.mimeType), self.mimeType);
-        try ipc.write(@TypeOf(self.isStream), self.isStream);
-        try ipc.write(@TypeOf(self.audioSystemSampleRate), self.audioSystemSampleRate);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureLoad {
-        var self: VideoTextureLoad = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.source = try ipc.read(@TypeOf(self.source));
-        self.overrideEngine = try ipc.read(@TypeOf(self.overrideEngine));
-        self.mimeType = try ipc.read(@TypeOf(self.mimeType));
-        self.isStream = try ipc.read(@TypeOf(self.isStream));
-        self.audioSystemSampleRate = try ipc.read(@TypeOf(self.audioSystemSampleRate));
-        return self;
-    }
-};
-
-pub const VideoTextureProperties = struct {
-    filterMode: TextureFilterMode,
-    anisoLevel: i32,
-    wrapU: TextureWrapMode,
-    wrapV: TextureWrapMode,
-    assetId: i32,
-
-    pub fn write(self: VideoTextureProperties, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.filterMode), self.filterMode);
-        try ipc.write(@TypeOf(self.anisoLevel), self.anisoLevel);
-        try ipc.write(@TypeOf(self.wrapU), self.wrapU);
-        try ipc.write(@TypeOf(self.wrapV), self.wrapV);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureProperties {
-        var self: VideoTextureProperties = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.filterMode = try ipc.read(@TypeOf(self.filterMode));
-        self.anisoLevel = try ipc.read(@TypeOf(self.anisoLevel));
-        self.wrapU = try ipc.read(@TypeOf(self.wrapU));
-        self.wrapV = try ipc.read(@TypeOf(self.wrapV));
-        return self;
-    }
-};
-
-pub const VideoTextureReady = struct {
-    length: f64,
-    size: math.Vector2i,
-    hasAlpha: bool,
-    playbackEngine: []const u16,
-    instanceChanged: bool,
-    audioTracks: []const VideoAudioTrack,
-    assetId: i32,
-
-    pub fn write(self: VideoTextureReady, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.length), self.length);
-        try ipc.write(@TypeOf(self.size), self.size);
-        try ipc.write(@TypeOf(self.hasAlpha), self.hasAlpha);
-        try ipc.write(@TypeOf(self.playbackEngine), self.playbackEngine);
-        try ipc.write(@TypeOf(self.instanceChanged), self.instanceChanged);
-        try ipc.writeList(@TypeOf(self.audioTracks), self.audioTracks);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureReady {
-        var self: VideoTextureReady = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.length = try ipc.read(@TypeOf(self.length));
-        self.size = try ipc.read(@TypeOf(self.size));
-        self.hasAlpha = try ipc.read(@TypeOf(self.hasAlpha));
-        self.playbackEngine = try ipc.read(@TypeOf(self.playbackEngine));
-        self.instanceChanged = try ipc.read(@TypeOf(self.instanceChanged));
-        self.audioTracks = try ipc.readList(@TypeOf(self.audioTracks));
-        return self;
-    }
-};
-
-pub const VideoTextureStartAudioTrack = struct {
-    audioTrackIndex: i32,
-    queueCapacity: i32,
-    queueName: []const u16,
-    assetId: i32,
-
-    pub fn write(self: VideoTextureStartAudioTrack, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.audioTrackIndex), self.audioTrackIndex);
-        try ipc.write(@TypeOf(self.queueCapacity), self.queueCapacity);
-        try ipc.write(@TypeOf(self.queueName), self.queueName);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureStartAudioTrack {
-        var self: VideoTextureStartAudioTrack = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.audioTrackIndex = try ipc.read(@TypeOf(self.audioTrackIndex));
-        self.queueCapacity = try ipc.read(@TypeOf(self.queueCapacity));
-        self.queueName = try ipc.read(@TypeOf(self.queueName));
-        return self;
-    }
-};
-
-pub const VideoTextureUpdate = struct {
-    position: f64,
-    play: bool,
-    loop: bool,
-    assetId: i32,
-
-    pub fn write(self: VideoTextureUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.assetId), self.assetId);
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write8PackedBools(self.play, self.loop, false, false, false, false, false, false);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VideoTextureUpdate {
-        var self: VideoTextureUpdate = undefined;
-        self.assetId = try ipc.read(@TypeOf(self.assetId));
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.play, self.loop, _, _, _, _, _, _ = try ipc.read8PackedBools();
-        return self;
-    }
-};
-
-pub const DesktopConfig = struct {
-    maximumBackgroundFramerate: ?i32,
-    maximumForegroundFramerate: ?i32,
-    vSync: bool,
-
-    pub fn write(self: DesktopConfig, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.maximumBackgroundFramerate), self.maximumBackgroundFramerate);
-        try ipc.write(@TypeOf(self.maximumBackgroundFramerate), self.maximumBackgroundFramerate);
-        try ipc.write(@TypeOf(self.vSync), self.vSync);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !DesktopConfig {
-        var self: DesktopConfig = undefined;
-        self.maximumBackgroundFramerate = try ipc.read(@TypeOf(self.maximumBackgroundFramerate));
-        self.maximumBackgroundFramerate = try ipc.read(@TypeOf(self.maximumBackgroundFramerate));
-        self.vSync = try ipc.read(@TypeOf(self.vSync));
-        return self;
-    }
-};
-
-pub const GaussianSplatConfig = struct {
-    sortingMegaOperationsPerCamera: f32,
-
-    pub fn write(self: GaussianSplatConfig, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.sortingMegaOperationsPerCamera), self.sortingMegaOperationsPerCamera);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GaussianSplatConfig {
-        var self: GaussianSplatConfig = undefined;
-        self.sortingMegaOperationsPerCamera = try ipc.read(@TypeOf(self.sortingMegaOperationsPerCamera));
-        return self;
-    }
-};
-
-pub const PostProcessingConfig = struct {
-    motionBlurIntensity: f32,
-    bloomIntensity: f32,
-    ambientOcclusionIntensity: f32,
-    screenSpaceReflections: bool,
-    antialiasing: AntiAliasingMethod,
-
-    pub fn write(self: PostProcessingConfig, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.motionBlurIntensity), self.motionBlurIntensity);
-        try ipc.write(@TypeOf(self.bloomIntensity), self.bloomIntensity);
-        try ipc.write(@TypeOf(self.ambientOcclusionIntensity), self.ambientOcclusionIntensity);
-        try ipc.write(@TypeOf(self.screenSpaceReflections), self.screenSpaceReflections);
-        try ipc.write(@TypeOf(self.antialiasing), self.antialiasing);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PostProcessingConfig {
-        var self: PostProcessingConfig = undefined;
-        self.motionBlurIntensity = try ipc.read(@TypeOf(self.motionBlurIntensity));
-        self.bloomIntensity = try ipc.read(@TypeOf(self.bloomIntensity));
-        self.ambientOcclusionIntensity = try ipc.read(@TypeOf(self.ambientOcclusionIntensity));
-        self.screenSpaceReflections = try ipc.read(@TypeOf(self.screenSpaceReflections));
-        self.antialiasing = try ipc.read(@TypeOf(self.antialiasing));
-        return self;
-    }
-};
-
-pub const QualityConfig = struct {
-    perPixelLight: i32,
-    shadowCascades: ShadowCascadeMode,
-    shadowResolution: ShadowResolutionMode,
-    shadowDistance: f32,
-    skinWeightMode: SkinWeightMode,
-
-    pub fn write(self: QualityConfig, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.perPixelLight), self.perPixelLight);
-        try ipc.write(@TypeOf(self.shadowCascades), self.shadowCascades);
-        try ipc.write(@TypeOf(self.shadowResolution), self.shadowResolution);
-        try ipc.write(@TypeOf(self.shadowDistance), self.shadowDistance);
-        try ipc.write(@TypeOf(self.skinWeightMode), self.skinWeightMode);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !QualityConfig {
-        var self: QualityConfig = undefined;
-        self.perPixelLight = try ipc.read(@TypeOf(self.perPixelLight));
-        self.shadowCascades = try ipc.read(@TypeOf(self.shadowCascades));
-        self.shadowResolution = try ipc.read(@TypeOf(self.shadowResolution));
-        self.shadowDistance = try ipc.read(@TypeOf(self.shadowDistance));
-        self.skinWeightMode = try ipc.read(@TypeOf(self.skinWeightMode));
-        return self;
-    }
-};
-
-pub const ResolutionConfig = struct {
-    resolution: math.Vector2i,
-    fullscreen: bool,
-
-    pub fn write(self: ResolutionConfig, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.resolution), self.resolution);
-        try ipc.write(@TypeOf(self.fullscreen), self.fullscreen);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ResolutionConfig {
-        var self: ResolutionConfig = undefined;
-        self.resolution = try ipc.read(@TypeOf(self.resolution));
-        self.fullscreen = try ipc.read(@TypeOf(self.fullscreen));
-        return self;
-    }
-};
-
-pub const CameraRenderParameters = struct {
-    resolution: math.Vector2i,
-    textureFormat: TextureFormat,
-    projection: CameraProjection,
-    fov: f32,
-    orthographicSize: f32,
-    clearMode: CameraClearMode,
-    clearColor: math.Vector4f,
-    nearClip: f32,
-    farClip: f32,
-    renderPrivateUI: bool,
-    postProcessing: bool,
-    screenSpaceReflections: bool,
-
-    pub fn write(self: CameraRenderParameters, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.resolution), self.resolution);
-        try ipc.write(@TypeOf(self.textureFormat), self.textureFormat);
-        try ipc.write(@TypeOf(self.projection), self.projection);
-        try ipc.write(@TypeOf(self.fov), self.fov);
-        try ipc.write(@TypeOf(self.orthographicSize), self.orthographicSize);
-        try ipc.write(@TypeOf(self.clearMode), self.clearMode);
-        try ipc.write(@TypeOf(self.clearColor), self.clearColor);
-        try ipc.write(@TypeOf(self.nearClip), self.nearClip);
-        try ipc.write(@TypeOf(self.farClip), self.farClip);
-        try ipc.write(@TypeOf(self.renderPrivateUI), self.renderPrivateUI);
-        try ipc.write(@TypeOf(self.postProcessing), self.postProcessing);
-        try ipc.write(@TypeOf(self.screenSpaceReflections), self.screenSpaceReflections);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraRenderParameters {
-        var self: CameraRenderParameters = undefined;
-        self.resolution = try ipc.read(@TypeOf(self.resolution));
-        self.textureFormat = try ipc.read(@TypeOf(self.textureFormat));
-        self.projection = try ipc.read(@TypeOf(self.projection));
-        self.fov = try ipc.read(@TypeOf(self.fov));
-        self.orthographicSize = try ipc.read(@TypeOf(self.orthographicSize));
-        self.clearMode = try ipc.read(@TypeOf(self.clearMode));
-        self.clearColor = try ipc.read(@TypeOf(self.clearColor));
-        self.nearClip = try ipc.read(@TypeOf(self.nearClip));
-        self.farClip = try ipc.read(@TypeOf(self.farClip));
-        self.renderPrivateUI = try ipc.read(@TypeOf(self.renderPrivateUI));
-        self.postProcessing = try ipc.read(@TypeOf(self.postProcessing));
-        self.screenSpaceReflections = try ipc.read(@TypeOf(self.screenSpaceReflections));
-        return self;
-    }
-};
-
-pub const CameraRenderTask = struct {
-    renderSpaceId: i32,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    parameters: ?CameraRenderParameters,
-    resultData: SharedMemoryBufferDescriptor,
-    onlyRenderList: []const i32,
-    excludeRenderList: []const i32,
-
-    pub fn write(self: CameraRenderTask, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderSpaceId), self.renderSpaceId);
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        try ipc.write(@TypeOf(self.parameters), self.parameters);
-        try ipc.write(@TypeOf(self.resultData), self.resultData);
-        try ipc.writeList(@TypeOf(self.onlyRenderList), self.onlyRenderList);
-        try ipc.writeList(@TypeOf(self.excludeRenderList), self.excludeRenderList);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraRenderTask {
-        var self: CameraRenderTask = undefined;
-        self.renderSpaceId = try ipc.read(@TypeOf(self.renderSpaceId));
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        self.parameters = try ipc.read(@TypeOf(self.parameters));
-        self.resultData = try ipc.read(@TypeOf(self.resultData));
-        self.onlyRenderList = try ipc.readList(@TypeOf(self.onlyRenderList));
-        self.excludeRenderList = try ipc.readList(@TypeOf(self.excludeRenderList));
-        return self;
-    }
-};
-
-pub const FrameStartData = struct {
-    lastFrameIndex: i32,
-    performance: ?PerformanceState,
-    inputs: ?InputState,
-    renderedReflectionProbes: []const RenderableHandle,
-
-    pub fn write(self: FrameStartData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.lastFrameIndex), self.lastFrameIndex);
-        try ipc.write(@TypeOf(self.performance), self.performance);
-        try ipc.write(@TypeOf(self.inputs), self.inputs);
-        try ipc.writeList(@TypeOf(self.renderedReflectionProbes), self.renderedReflectionProbes);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !FrameStartData {
-        var self: FrameStartData = undefined;
-        self.lastFrameIndex = try ipc.read(@TypeOf(self.lastFrameIndex));
-        self.performance = try ipc.read(@TypeOf(self.performance));
-        self.inputs = try ipc.read(@TypeOf(self.inputs));
-        self.renderedReflectionProbes = try ipc.readList(@TypeOf(self.renderedReflectionProbes));
-        return self;
-    }
-};
-
-pub const FrameSubmitData = struct {
-    frameIndex: i32,
-    debugLog: bool,
-    vrActive: bool,
-    nearClip: f32,
-    farClip: f32,
-    desktopFOV: f32,
-    outputState: ?OutputState,
-    renderSpaces: []const RenderSpaceUpdate,
-    renderTasks: []const CameraRenderTask,
-
-    pub fn write(self: FrameSubmitData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.frameIndex), self.frameIndex);
-        try ipc.write(@TypeOf(self.debugLog), self.debugLog);
-        try ipc.write(@TypeOf(self.vrActive), self.vrActive);
-        try ipc.write(@TypeOf(self.nearClip), self.nearClip);
-        try ipc.write(@TypeOf(self.farClip), self.farClip);
-        try ipc.write(@TypeOf(self.desktopFOV), self.desktopFOV);
-        try ipc.write(@TypeOf(self.outputState), self.outputState);
-        try ipc.writeList(@TypeOf(self.renderSpaces), self.renderSpaces);
-        try ipc.writeList(@TypeOf(self.renderTasks), self.renderTasks);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !FrameSubmitData {
-        var self: FrameSubmitData = undefined;
-        self.frameIndex = try ipc.read(@TypeOf(self.frameIndex));
-        self.debugLog = try ipc.read(@TypeOf(self.debugLog));
-        self.vrActive = try ipc.read(@TypeOf(self.vrActive));
-        self.nearClip = try ipc.read(@TypeOf(self.nearClip));
-        self.farClip = try ipc.read(@TypeOf(self.farClip));
-        self.desktopFOV = try ipc.read(@TypeOf(self.desktopFOV));
-        self.outputState = try ipc.read(@TypeOf(self.outputState));
-        self.renderSpaces = try ipc.readList(@TypeOf(self.renderSpaces));
-        self.renderTasks = try ipc.readList(@TypeOf(self.renderTasks));
-        return self;
-    }
-};
-
-pub const KeepAlive = struct {};
-
-pub const PerformanceState = struct {
-    fps: f32,
-    immediateFPS: f32,
-    renderTime: f32,
-    externalUpdateTime: f32,
-    frameBeginToSubmitTime: f32,
-    frameProcessedToNextBeginTime: f32,
-    integrationProcessingTime: f32,
-    extraParticleProcessingTime: f32,
-    processedAssetIntegratorTasks: i32,
-    integrationHighPriorityTasks: i32,
-    integrationTasks: i32,
-    integrationRenderTasks: i32,
-    integrationParticleTasks: i32,
-    processingHandleWaits: i32,
-    frameUpdateHandleTime: f32,
-    renderedCameras: i32,
-    renderedCameraPortals: i32,
-    updatedTextures: i32,
-    textureSliceUploads: i32,
-
-    pub fn write(self: PerformanceState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.fps), self.fps);
-        try ipc.write(@TypeOf(self.immediateFPS), self.immediateFPS);
-        try ipc.write(@TypeOf(self.renderTime), self.renderTime);
-        try ipc.write(@TypeOf(self.externalUpdateTime), self.externalUpdateTime);
-        try ipc.write(@TypeOf(self.frameBeginToSubmitTime), self.frameBeginToSubmitTime);
-        try ipc.write(@TypeOf(self.frameProcessedToNextBeginTime), self.frameProcessedToNextBeginTime);
-        try ipc.write(@TypeOf(self.integrationProcessingTime), self.integrationProcessingTime);
-        try ipc.write(@TypeOf(self.extraParticleProcessingTime), self.extraParticleProcessingTime);
-        try ipc.write(@TypeOf(self.processedAssetIntegratorTasks), self.processedAssetIntegratorTasks);
-        try ipc.write(@TypeOf(self.integrationHighPriorityTasks), self.integrationHighPriorityTasks);
-        try ipc.write(@TypeOf(self.integrationTasks), self.integrationTasks);
-        try ipc.write(@TypeOf(self.integrationRenderTasks), self.integrationRenderTasks);
-        try ipc.write(@TypeOf(self.integrationParticleTasks), self.integrationParticleTasks);
-        try ipc.write(@TypeOf(self.processingHandleWaits), self.processingHandleWaits);
-        try ipc.write(@TypeOf(self.frameUpdateHandleTime), self.frameUpdateHandleTime);
-        try ipc.write(@TypeOf(self.renderedCameras), self.renderedCameras);
-        try ipc.write(@TypeOf(self.renderedCameraPortals), self.renderedCameraPortals);
-        try ipc.write(@TypeOf(self.updatedTextures), self.updatedTextures);
-        try ipc.write(@TypeOf(self.textureSliceUploads), self.textureSliceUploads);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PerformanceState {
-        var self: PerformanceState = undefined;
-        self.fps = try ipc.read(@TypeOf(self.fps));
-        self.immediateFPS = try ipc.read(@TypeOf(self.immediateFPS));
-        self.renderTime = try ipc.read(@TypeOf(self.renderTime));
-        self.externalUpdateTime = try ipc.read(@TypeOf(self.externalUpdateTime));
-        self.frameBeginToSubmitTime = try ipc.read(@TypeOf(self.frameBeginToSubmitTime));
-        self.frameProcessedToNextBeginTime = try ipc.read(@TypeOf(self.frameProcessedToNextBeginTime));
-        self.integrationProcessingTime = try ipc.read(@TypeOf(self.integrationProcessingTime));
-        self.extraParticleProcessingTime = try ipc.read(@TypeOf(self.extraParticleProcessingTime));
-        self.processedAssetIntegratorTasks = try ipc.read(@TypeOf(self.processedAssetIntegratorTasks));
-        self.integrationHighPriorityTasks = try ipc.read(@TypeOf(self.integrationHighPriorityTasks));
-        self.integrationTasks = try ipc.read(@TypeOf(self.integrationTasks));
-        self.integrationRenderTasks = try ipc.read(@TypeOf(self.integrationRenderTasks));
-        self.integrationParticleTasks = try ipc.read(@TypeOf(self.integrationParticleTasks));
-        self.processingHandleWaits = try ipc.read(@TypeOf(self.processingHandleWaits));
-        self.frameUpdateHandleTime = try ipc.read(@TypeOf(self.frameUpdateHandleTime));
-        self.renderedCameras = try ipc.read(@TypeOf(self.renderedCameras));
-        self.renderedCameraPortals = try ipc.read(@TypeOf(self.renderedCameraPortals));
-        self.updatedTextures = try ipc.read(@TypeOf(self.updatedTextures));
-        self.textureSliceUploads = try ipc.read(@TypeOf(self.textureSliceUploads));
-        return self;
-    }
-};
-
-pub const ReflectionProbeRenderResult = struct {
-    renderTaskId: i32,
-    success: bool,
-
-    pub fn write(self: ReflectionProbeRenderResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderTaskId), self.renderTaskId);
-        try ipc.write(@TypeOf(self.success), self.success);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ReflectionProbeRenderResult {
-        var self: ReflectionProbeRenderResult = undefined;
-        self.renderTaskId = try ipc.read(@TypeOf(self.renderTaskId));
-        self.success = try ipc.read(@TypeOf(self.success));
-        return self;
-    }
-};
-
-pub const ReflectionProbeRenderTask = struct {
-    renderableIndex: i32,
-    renderTaskId: i32,
-    size: i32,
-    hdr: bool,
-    mipOrigins: []const []const i32,
-    resultData: SharedMemoryBufferDescriptor,
-    excludeTransformIds: []const i32,
-
-    pub fn write(self: ReflectionProbeRenderTask, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.renderTaskId), self.renderTaskId);
-        try ipc.write(@TypeOf(self.size), self.size);
-        try ipc.write(@TypeOf(self.hdr), self.hdr);
-        try ipc.writeNestedList(@TypeOf(self.mipOrigins), self.mipOrigins);
-        try ipc.write(@TypeOf(self.resultData), self.resultData);
-        try ipc.writeList(@TypeOf(self.excludeTransformIds), self.excludeTransformIds);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ReflectionProbeRenderTask {
-        var self: ReflectionProbeRenderTask = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.renderTaskId = try ipc.read(@TypeOf(self.renderTaskId));
-        self.size = try ipc.read(@TypeOf(self.size));
-        self.hdr = try ipc.read(@TypeOf(self.hdr));
-        self.mipOrigins = try ipc.readNestedList(@TypeOf(self.mipOrigins));
-        self.resultData = try ipc.read(@TypeOf(self.resultData));
-        self.excludeTransformIds = try ipc.readList(@TypeOf(self.excludeTransformIds));
-        return self;
-    }
-};
-
-pub const RenderSpaceUpdate = struct {
-    id: i32,
-    isActive: bool,
-    isOverlay: bool,
-    isPrivate: bool,
-    rootTransform: RenderTransform,
-    viewPositionIsExternal: bool,
-    overrideViewPosition: bool,
-    skyboxMaterialAssetId: i32,
-    ambientLight: RenderSH2,
-    overridenViewTransform: RenderTransform,
-    transformsUpdate: ?TransformsUpdate,
-    meshRenderersUpdate: ?MeshRenderablesUpdate,
-    skinnedMeshRenderersUpdate: ?SkinnedMeshRenderablesUpdate,
-    lightsUpdate: ?LightRenderablesUpdate,
-    camerasUpdate: ?CameraRenderablesUpdate,
-    cameraPortalsUpdate: ?CameraPortalsRenderablesUpdate,
-    reflectionProbesUpdate: ?ReflectionProbeRenderablesUpdate,
-    reflectionProbeSH2Taks: ?ReflectionProbeSH2Tasks,
-    layersUpdate: ?LayerUpdate,
-    billboardBuffersUpdate: ?BillboardRenderBufferUpdate,
-    meshRenderBuffersUpdate: ?MeshRenderBufferUpdate,
-    trailRenderersUpdate: ?TrailsRendererUpdate,
-    lightsBufferRenderersUpdate: ?LightsBufferRendererUpdate,
-    renderTransformOverridesUpdate: ?RenderTransformOverridesUpdate,
-    renderMaterialOverridesUpdate: ?RenderMaterialOverridesUpdate,
-    blitToDisplaysUpdate: ?BlitToDisplayRenderablesUpdate,
-    lodGroupUpdate: ?LODGroupRenderablesUpdate,
-    gaussianSplatRenderersUpdate: ?GaussianSplatRenderablesUpdate,
-    reflectionProbeRenderTasks: []const ReflectionProbeRenderTask,
-
-    pub fn write(self: RenderSpaceUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.id), self.id);
-        try ipc.write(@TypeOf(self.isActive), self.isActive);
-        try ipc.write(@TypeOf(self.isOverlay), self.isOverlay);
-        try ipc.write(@TypeOf(self.isPrivate), self.isPrivate);
-        try ipc.write(@TypeOf(self.rootTransform), self.rootTransform);
-        try ipc.write(@TypeOf(self.viewPositionIsExternal), self.viewPositionIsExternal);
-        try ipc.write(@TypeOf(self.overrideViewPosition), self.overrideViewPosition);
-        try ipc.write(@TypeOf(self.skyboxMaterialAssetId), self.skyboxMaterialAssetId);
-        try ipc.write(@TypeOf(self.ambientLight), self.ambientLight);
-        try ipc.write(@TypeOf(self.overridenViewTransform), self.overridenViewTransform);
-        try ipc.write(@TypeOf(self.transformsUpdate), self.transformsUpdate);
-        try ipc.write(@TypeOf(self.meshRenderersUpdate), self.meshRenderersUpdate);
-        try ipc.write(@TypeOf(self.skinnedMeshRenderersUpdate), self.skinnedMeshRenderersUpdate);
-        try ipc.write(@TypeOf(self.lightsUpdate), self.lightsUpdate);
-        try ipc.write(@TypeOf(self.camerasUpdate), self.camerasUpdate);
-        try ipc.write(@TypeOf(self.cameraPortalsUpdate), self.cameraPortalsUpdate);
-        try ipc.write(@TypeOf(self.reflectionProbesUpdate), self.reflectionProbesUpdate);
-        try ipc.write(@TypeOf(self.reflectionProbeSH2Taks), self.reflectionProbeSH2Taks);
-        try ipc.write(@TypeOf(self.layersUpdate), self.layersUpdate);
-        try ipc.write(@TypeOf(self.billboardBuffersUpdate), self.billboardBuffersUpdate);
-        try ipc.write(@TypeOf(self.meshRenderBuffersUpdate), self.meshRenderBuffersUpdate);
-        try ipc.write(@TypeOf(self.trailRenderersUpdate), self.trailRenderersUpdate);
-        try ipc.write(@TypeOf(self.lightsBufferRenderersUpdate), self.lightsBufferRenderersUpdate);
-        try ipc.write(@TypeOf(self.renderTransformOverridesUpdate), self.renderTransformOverridesUpdate);
-        try ipc.write(@TypeOf(self.renderMaterialOverridesUpdate), self.renderMaterialOverridesUpdate);
-        try ipc.write(@TypeOf(self.blitToDisplaysUpdate), self.blitToDisplaysUpdate);
-        try ipc.write(@TypeOf(self.lodGroupUpdate), self.lodGroupUpdate);
-        try ipc.write(@TypeOf(self.gaussianSplatRenderersUpdate), self.gaussianSplatRenderersUpdate);
-        try ipc.writeList(@TypeOf(self.reflectionProbeRenderTasks), self.reflectionProbeRenderTasks);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderSpaceUpdate {
-        var self: RenderSpaceUpdate = undefined;
-        self.id = try ipc.read(@TypeOf(self.id));
-        self.isActive = try ipc.read(@TypeOf(self.isActive));
-        self.isOverlay = try ipc.read(@TypeOf(self.isOverlay));
-        self.isPrivate = try ipc.read(@TypeOf(self.isPrivate));
-        self.rootTransform = try ipc.read(@TypeOf(self.rootTransform));
-        self.viewPositionIsExternal = try ipc.read(@TypeOf(self.viewPositionIsExternal));
-        self.overrideViewPosition = try ipc.read(@TypeOf(self.overrideViewPosition));
-        self.skyboxMaterialAssetId = try ipc.read(@TypeOf(self.skyboxMaterialAssetId));
-        self.ambientLight = try ipc.read(@TypeOf(self.ambientLight));
-        self.overridenViewTransform = try ipc.read(@TypeOf(self.overridenViewTransform));
-        self.transformsUpdate = try ipc.read(@TypeOf(self.transformsUpdate));
-        self.meshRenderersUpdate = try ipc.read(@TypeOf(self.meshRenderersUpdate));
-        self.skinnedMeshRenderersUpdate = try ipc.read(@TypeOf(self.skinnedMeshRenderersUpdate));
-        self.lightsUpdate = try ipc.read(@TypeOf(self.lightsUpdate));
-        self.camerasUpdate = try ipc.read(@TypeOf(self.camerasUpdate));
-        self.cameraPortalsUpdate = try ipc.read(@TypeOf(self.cameraPortalsUpdate));
-        self.reflectionProbesUpdate = try ipc.read(@TypeOf(self.reflectionProbesUpdate));
-        self.reflectionProbeSH2Taks = try ipc.read(@TypeOf(self.reflectionProbeSH2Taks));
-        self.layersUpdate = try ipc.read(@TypeOf(self.layersUpdate));
-        self.billboardBuffersUpdate = try ipc.read(@TypeOf(self.billboardBuffersUpdate));
-        self.meshRenderBuffersUpdate = try ipc.read(@TypeOf(self.meshRenderBuffersUpdate));
-        self.trailRenderersUpdate = try ipc.read(@TypeOf(self.trailRenderersUpdate));
-        self.lightsBufferRenderersUpdate = try ipc.read(@TypeOf(self.lightsBufferRenderersUpdate));
-        self.renderTransformOverridesUpdate = try ipc.read(@TypeOf(self.renderTransformOverridesUpdate));
-        self.renderMaterialOverridesUpdate = try ipc.read(@TypeOf(self.renderMaterialOverridesUpdate));
-        self.blitToDisplaysUpdate = try ipc.read(@TypeOf(self.blitToDisplaysUpdate));
-        self.lodGroupUpdate = try ipc.read(@TypeOf(self.lodGroupUpdate));
-        self.gaussianSplatRenderersUpdate = try ipc.read(@TypeOf(self.gaussianSplatRenderersUpdate));
-        self.reflectionProbeRenderTasks = try ipc.readList(@TypeOf(self.reflectionProbeRenderTasks));
-        return self;
-    }
-};
-
-pub const ViveHandState = struct {
-    confidence: f32,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    pinchStrength: f32,
-    points: []const math.Vector3f,
-
-    pub fn write(self: ViveHandState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.confidence), self.confidence);
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        try ipc.write(@TypeOf(self.pinchStrength), self.pinchStrength);
-        try ipc.writeList(@TypeOf(self.points), self.points);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ViveHandState {
-        var self: ViveHandState = undefined;
-        self.confidence = try ipc.read(@TypeOf(self.confidence));
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        self.pinchStrength = try ipc.read(@TypeOf(self.pinchStrength));
-        self.points = try ipc.readList(@TypeOf(self.points));
-        return self;
-    }
-};
-
-pub const ViveHandTrackingInputState = struct {
-    isTracking: bool,
-    left: ?ViveHandState,
-    right: ?ViveHandState,
-
-    pub fn write(self: ViveHandTrackingInputState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.isTracking), self.isTracking);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0012: brfalse.s IL_002c
-        try ipc.write(@TypeOf(self.left), self.left);
-        try ipc.write(@TypeOf(self.right), self.right);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ViveHandTrackingInputState {
-        var self: ViveHandTrackingInputState = undefined;
-        self.isTracking = try ipc.read(@TypeOf(self.isTracking));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0012: brfalse.s IL_002c
-        self.left = try ipc.read(@TypeOf(self.left));
-        self.right = try ipc.read(@TypeOf(self.right));
-        return self;
-    }
-};
-
-pub const DisplayState = struct {
-    displayIndex: i32,
-    resolution: math.Vector2i,
-    offset: math.Vector2i,
-    refreshRate: f64,
-    orientation: RectOrientation,
-    dpi: math.Vector2f,
-    isPrimary: bool,
-
-    pub fn write(self: DisplayState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.displayIndex), self.displayIndex);
-        try ipc.write(@TypeOf(self.resolution), self.resolution);
-        try ipc.write(@TypeOf(self.offset), self.offset);
-        try ipc.write(@TypeOf(self.refreshRate), self.refreshRate);
-        try ipc.write(@TypeOf(self.orientation), self.orientation);
-        try ipc.write(@TypeOf(self.dpi), self.dpi);
-        try ipc.write(@TypeOf(self.isPrimary), self.isPrimary);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !DisplayState {
-        var self: DisplayState = undefined;
-        self.displayIndex = try ipc.read(@TypeOf(self.displayIndex));
-        self.resolution = try ipc.read(@TypeOf(self.resolution));
-        self.offset = try ipc.read(@TypeOf(self.offset));
-        self.refreshRate = try ipc.read(@TypeOf(self.refreshRate));
-        self.orientation = try ipc.read(@TypeOf(self.orientation));
-        self.dpi = try ipc.read(@TypeOf(self.dpi));
-        self.isPrimary = try ipc.read(@TypeOf(self.isPrimary));
-        return self;
-    }
-};
-
-pub const DragAndDropEvent = struct {
-    paths: []const []const u16,
-    dropPoint: math.Vector2i,
-
-    pub fn write(self: DragAndDropEvent, ipc: IpcSerializer) !void {
-        try ipc.writeList(@TypeOf(self.paths), self.paths);
-        try ipc.write(@TypeOf(self.dropPoint), self.dropPoint);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !DragAndDropEvent {
-        var self: DragAndDropEvent = undefined;
-        self.paths = try ipc.readList(@TypeOf(self.paths));
-        self.dropPoint = try ipc.read(@TypeOf(self.dropPoint));
-        return self;
-    }
-};
-
-pub const GamepadState = struct {
-    displayName: []const u16,
-    leftThumbstick: math.Vector2f,
-    rightThumbstick: math.Vector2f,
-    dPad: math.Vector2f,
-    leftTrigger: f32,
-    rightTrigger: f32,
-    leftThumbstickClick: bool,
-    rightThumbstickClick: bool,
-    dPadUp: bool,
-    dPadRight: bool,
-    dPadDown: bool,
-    dPadLeft: bool,
-    leftBumper: bool,
-    rightBumper: bool,
-    start: bool,
-    menu: bool,
-    a: bool,
-    b: bool,
-    x: bool,
-    y: bool,
-    paddle0: bool,
-    paddle1: bool,
-    paddle2: bool,
-    paddle3: bool,
-
-    pub fn write(self: GamepadState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.displayName), self.displayName);
-        try ipc.write(@TypeOf(self.leftThumbstick), self.leftThumbstick);
-        try ipc.write(@TypeOf(self.rightThumbstick), self.rightThumbstick);
-        try ipc.write(@TypeOf(self.dPad), self.dPad);
-        try ipc.write(@TypeOf(self.leftTrigger), self.leftTrigger);
-        try ipc.write(@TypeOf(self.rightTrigger), self.rightTrigger);
-        try ipc.write8PackedBools(self.leftThumbstickClick, self.rightThumbstickClick, self.dPadUp, self.dPadRight, self.dPadDown, self.dPadLeft, self.leftBumper, self.rightBumper);
-        try ipc.write8PackedBools(self.start, self.menu, false, false, false, false, false, false);
-        try ipc.write8PackedBools(self.a, self.b, self.x, self.y, self.paddle0, self.paddle1, self.paddle2, self.paddle3);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GamepadState {
-        var self: GamepadState = undefined;
-        self.displayName = try ipc.read(@TypeOf(self.displayName));
-        self.leftThumbstick = try ipc.read(@TypeOf(self.leftThumbstick));
-        self.rightThumbstick = try ipc.read(@TypeOf(self.rightThumbstick));
-        self.dPad = try ipc.read(@TypeOf(self.dPad));
-        self.leftTrigger = try ipc.read(@TypeOf(self.leftTrigger));
-        self.rightTrigger = try ipc.read(@TypeOf(self.rightTrigger));
-        self.leftThumbstickClick, self.rightThumbstickClick, self.dPadUp, self.dPadRight, self.dPadDown, self.dPadLeft, self.leftBumper, self.rightBumper = try ipc.read8PackedBools();
-        self.start, self.menu, _, _, _, _, _, _ = try ipc.read8PackedBools();
-        self.a, self.b, self.x, self.y, self.paddle0, self.paddle1, self.paddle2, self.paddle3 = try ipc.read8PackedBools();
-        return self;
-    }
-};
-
-pub const HandState = struct {
-    uniqueId: []const u16,
-    priority: i32,
-    chirality: Chirality,
-    isDeviceActive: bool,
-    isTracking: bool,
-    tracksMetacarpals: bool,
-    confidence: f32,
-    wristPosition: math.Vector3f,
-    wristRotation: math.Quaternionf,
-    segmentPositions: []const math.Vector3f,
-    segmentRotations: []const math.Quaternionf,
-
-    pub fn write(self: HandState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.uniqueId), self.uniqueId);
-        try ipc.write(@TypeOf(self.priority), self.priority);
-        try ipc.write(@TypeOf(self.chirality), self.chirality);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.tracksMetacarpals, false, false, false, false, false);
-        try ipc.write(@TypeOf(self.confidence), self.confidence);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_0085
-        try ipc.write(@TypeOf(self.wristPosition), self.wristPosition);
-        try ipc.write(@TypeOf(self.wristRotation), self.wristRotation);
-        try ipc.writeList(@TypeOf(self.segmentPositions), self.segmentPositions);
-        try ipc.writeList(@TypeOf(self.segmentRotations), self.segmentRotations);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !HandState {
-        var self: HandState = undefined;
-        self.uniqueId = try ipc.read(@TypeOf(self.uniqueId));
-        self.priority = try ipc.read(@TypeOf(self.priority));
-        self.chirality = try ipc.read(@TypeOf(self.chirality));
-        self.isDeviceActive, self.isTracking, self.tracksMetacarpals, _, _, _, _, _ = try ipc.read8PackedBools();
-        self.confidence = try ipc.read(@TypeOf(self.confidence));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_0080
-        self.wristPosition = try ipc.read(@TypeOf(self.wristPosition));
-        self.wristRotation = try ipc.read(@TypeOf(self.wristRotation));
-        self.segmentPositions = try ipc.readList(@TypeOf(self.segmentPositions));
-        self.segmentRotations = try ipc.readList(@TypeOf(self.segmentRotations));
-        return self;
-    }
-};
-
-pub const InputState = struct {
-    mouse: ?MouseState,
-    keyboard: ?KeyboardState,
-    window: ?WindowState,
-    vr: ?VR_InputsState,
-    gamepads: []const GamepadState,
-    touches: []const TouchState,
-    displays: []const DisplayState,
-
-    pub fn write(self: InputState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.mouse), self.mouse);
-        try ipc.write(@TypeOf(self.keyboard), self.keyboard);
-        try ipc.write(@TypeOf(self.window), self.window);
-        try ipc.write(@TypeOf(self.vr), self.vr);
-        try ipc.writeList(@TypeOf(self.gamepads), self.gamepads);
-        try ipc.writeList(@TypeOf(self.touches), self.touches);
-        try ipc.writeList(@TypeOf(self.displays), self.displays);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !InputState {
-        var self: InputState = undefined;
-        self.mouse = try ipc.read(@TypeOf(self.mouse));
-        self.keyboard = try ipc.read(@TypeOf(self.keyboard));
-        self.window = try ipc.read(@TypeOf(self.window));
-        self.vr = try ipc.read(@TypeOf(self.vr));
-        self.gamepads = try ipc.readList(@TypeOf(self.gamepads));
-        self.touches = try ipc.readList(@TypeOf(self.touches));
-        self.displays = try ipc.readList(@TypeOf(self.displays));
-        return self;
-    }
-};
-
-pub const KeyboardState = struct {
-    typeDelta: []const u16,
-    heldKeys: []const Key,
-
-    pub fn write(self: KeyboardState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.typeDelta), self.typeDelta);
-        try ipc.writeList(@TypeOf(self.heldKeys), self.heldKeys);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !KeyboardState {
-        var self: KeyboardState = undefined;
-        self.typeDelta = try ipc.read(@TypeOf(self.typeDelta));
-        self.heldKeys = try ipc.readList(@TypeOf(self.heldKeys));
-        return self;
-    }
-};
-
-pub const MouseState = struct {
-    isActive: bool,
-    leftButtonState: bool,
-    rightButtonState: bool,
-    middleButtonState: bool,
-    button4State: bool,
-    button5State: bool,
-    desktopPosition: math.Vector2f,
-    windowPosition: math.Vector2f,
-    directDelta: math.Vector2f,
-    scrollWheelDelta: math.Vector2f,
-
-    pub fn write(self: MouseState, ipc: IpcSerializer) !void {
-        try ipc.write8PackedBools(self.isActive, self.leftButtonState, self.rightButtonState, self.middleButtonState, self.button4State, self.button5State, false, false);
-        try ipc.write(@TypeOf(self.desktopPosition), self.desktopPosition);
-        try ipc.write(@TypeOf(self.windowPosition), self.windowPosition);
-        try ipc.write(@TypeOf(self.directDelta), self.directDelta);
-        try ipc.write(@TypeOf(self.scrollWheelDelta), self.scrollWheelDelta);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MouseState {
-        var self: MouseState = undefined;
-        self.isActive, self.leftButtonState, self.rightButtonState, self.middleButtonState, self.button4State, self.button5State, _, _ = try ipc.read8PackedBools();
-        self.desktopPosition = try ipc.read(@TypeOf(self.desktopPosition));
-        self.windowPosition = try ipc.read(@TypeOf(self.windowPosition));
-        self.directDelta = try ipc.read(@TypeOf(self.directDelta));
-        self.scrollWheelDelta = try ipc.read(@TypeOf(self.scrollWheelDelta));
-        return self;
-    }
-};
-
-pub const TouchState = struct {
-    touchId: i32,
-    position: math.Vector2f,
-    isPressing: bool,
-    pressure: f32,
-
-    pub fn write(self: TouchState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.touchId), self.touchId);
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.isPressing), self.isPressing);
-        try ipc.write(@TypeOf(self.pressure), self.pressure);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TouchState {
-        var self: TouchState = undefined;
-        self.touchId = try ipc.read(@TypeOf(self.touchId));
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.isPressing = try ipc.read(@TypeOf(self.isPressing));
-        self.pressure = try ipc.read(@TypeOf(self.pressure));
-        return self;
-    }
-};
-
-pub const CosmosControllerState = struct {
-    joystickTouch: bool,
-    joystickClick: bool,
-    joystickRaw: math.Vector2f,
-    triggerTouch: bool,
-    triggerClick: bool,
-    trigger: f32,
-    gripClick: bool,
-    vive: bool,
-    buttonAX: bool,
-    buttonBY: bool,
-    bumper: bool,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: CosmosControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.joystickTouch, self.joystickClick, self.triggerTouch, self.triggerClick, self.gripClick, self.vive, self.buttonAX, self.buttonBY);
-        try ipc.write(@TypeOf(self.joystickRaw), self.joystickRaw);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-        try ipc.write(@TypeOf(self.bumper), self.bumper);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CosmosControllerState {
-        var self: CosmosControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.joystickTouch, self.joystickClick, self.triggerTouch, self.triggerClick, self.gripClick, self.vive, self.buttonAX, self.buttonBY = try ipc.read8PackedBools();
-        self.joystickRaw = try ipc.read(@TypeOf(self.joystickRaw));
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        self.bumper = try ipc.read(@TypeOf(self.bumper));
-        return self;
-    }
-};
-
-pub const GenericControllerState = struct {
-    strength: f32,
-    axis: math.Vector2f,
-    touchingStrength: bool,
-    touchingAxis: bool,
-    primary: bool,
-    menu: bool,
-    grab: bool,
-    secondary: bool,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: GenericControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write(@TypeOf(self.strength), self.strength);
-        try ipc.write(@TypeOf(self.axis), self.axis);
-        try ipc.write8PackedBools(self.touchingStrength, self.touchingAxis, self.primary, self.menu, self.grab, self.secondary, false, false);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GenericControllerState {
-        var self: GenericControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.strength = try ipc.read(@TypeOf(self.strength));
-        self.axis = try ipc.read(@TypeOf(self.axis));
-        self.touchingStrength, self.touchingAxis, self.primary, self.menu, self.grab, self.secondary, _, _ = try ipc.read8PackedBools();
-        return self;
-    }
-};
-
-pub const HP_ReverbControllerState = struct {
-    appMenu: bool,
-    buttonYB: bool,
-    buttonXA: bool,
-    gripTouch: bool,
-    gripClick: bool,
-    grip: f32,
-    joystickClick: bool,
-    joystickRaw: math.Vector2f,
-    triggerHair: bool,
-    triggerClick: bool,
-    trigger: f32,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: HP_ReverbControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.appMenu, self.buttonYB, self.buttonXA, self.gripTouch, self.gripClick, self.joystickClick, self.triggerHair, self.triggerClick);
-        try ipc.write(@TypeOf(self.grip), self.grip);
-        try ipc.write(@TypeOf(self.joystickRaw), self.joystickRaw);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !HP_ReverbControllerState {
-        var self: HP_ReverbControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.appMenu, self.buttonYB, self.buttonXA, self.gripTouch, self.gripClick, self.joystickClick, self.triggerHair, self.triggerClick = try ipc.read8PackedBools();
-        self.grip = try ipc.read(@TypeOf(self.grip));
-        self.joystickRaw = try ipc.read(@TypeOf(self.joystickRaw));
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        return self;
-    }
-};
-
-pub const IndexControllerState = struct {
-    grip: f32,
-    gripTouch: bool,
-    gripClick: bool,
-    buttonA: bool,
-    buttonB: bool,
-    buttonAtouch: bool,
-    buttonBtouch: bool,
-    trigger: f32,
-    triggerTouch: bool,
-    triggerClick: bool,
-    joystickRaw: math.Vector2f,
-    joystickTouch: bool,
-    joystickClick: bool,
-    touchpad: math.Vector2f,
-    touchpadTouch: bool,
-    touchpadPress: bool,
-    touchpadForce: f32,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: IndexControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.gripTouch, self.gripClick, self.buttonA, self.buttonB, self.buttonAtouch, self.buttonBtouch, self.triggerTouch, self.triggerClick);
-        try ipc.write(@TypeOf(self.grip), self.grip);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-        try ipc.write8PackedBools(self.joystickTouch, self.joystickClick, self.touchpadTouch, self.touchpadPress, false, false, false, false);
-        try ipc.write(@TypeOf(self.joystickRaw), self.joystickRaw);
-        try ipc.write(@TypeOf(self.touchpad), self.touchpad);
-        try ipc.write(@TypeOf(self.touchpadForce), self.touchpadForce);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !IndexControllerState {
-        var self: IndexControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.gripTouch, self.gripClick, self.buttonA, self.buttonB, self.buttonAtouch, self.buttonBtouch, self.triggerTouch, self.triggerClick = try ipc.read8PackedBools();
-        self.grip = try ipc.read(@TypeOf(self.grip));
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        self.joystickTouch, self.joystickClick, self.touchpadTouch, self.touchpadPress, _, _, _, _ = try ipc.read8PackedBools();
-        self.joystickRaw = try ipc.read(@TypeOf(self.joystickRaw));
-        self.touchpad = try ipc.read(@TypeOf(self.touchpad));
-        self.touchpadForce = try ipc.read(@TypeOf(self.touchpadForce));
-        return self;
-    }
-};
-
-pub const PicoNeo2ControllerState = struct {
-    app: bool,
-    pico: bool,
-    buttonYB: bool,
-    buttonXA: bool,
-    gripClick: bool,
-    joystickTouch: bool,
-    joystickClick: bool,
-    joystick: math.Vector2f,
-    triggerClick: bool,
-    trigger: f32,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: PicoNeo2ControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.app, self.pico, self.buttonYB, self.buttonXA, self.gripClick, self.joystickTouch, self.joystickClick, self.triggerClick);
-        try ipc.write(@TypeOf(self.joystick), self.joystick);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !PicoNeo2ControllerState {
-        var self: PicoNeo2ControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.app, self.pico, self.buttonYB, self.buttonXA, self.gripClick, self.joystickTouch, self.joystickClick, self.triggerClick = try ipc.read8PackedBools();
-        self.joystick = try ipc.read(@TypeOf(self.joystick));
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        return self;
-    }
-};
-
-pub const TouchControllerState = struct {
-    model: TouchControllerModel,
-    start: bool,
-    buttonYB: bool,
-    buttonXA: bool,
-    buttonYB_touch: bool,
-    buttonXA_touch: bool,
-    thumbrestTouch: bool,
-    grip: f32,
-    gripClick: bool,
-    joystickRaw: math.Vector2f,
-    joystickTouch: bool,
-    joystickClick: bool,
-    trigger: f32,
-    triggerTouch: bool,
-    triggerClick: bool,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: TouchControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write(@TypeOf(self.model), self.model);
-        try ipc.write8PackedBools(self.start, self.buttonYB, self.buttonXA, self.buttonYB_touch, self.buttonXA_touch, self.thumbrestTouch, false, false);
-        try ipc.write8PackedBools(self.gripClick, self.joystickTouch, self.joystickClick, self.triggerTouch, self.triggerClick, false, false, false);
-        try ipc.write(@TypeOf(self.grip), self.grip);
-        try ipc.write(@TypeOf(self.joystickRaw), self.joystickRaw);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TouchControllerState {
-        var self: TouchControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.model = try ipc.read(@TypeOf(self.model));
-        self.start, self.buttonYB, self.buttonXA, self.buttonYB_touch, self.buttonXA_touch, self.thumbrestTouch, _, _ = try ipc.read8PackedBools();
-        self.gripClick, self.joystickTouch, self.joystickClick, self.triggerTouch, self.triggerClick, _, _, _ = try ipc.read8PackedBools();
-        self.grip = try ipc.read(@TypeOf(self.grip));
-        self.joystickRaw = try ipc.read(@TypeOf(self.joystickRaw));
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        return self;
-    }
-};
-
-pub const ViveControllerState = struct {
-    grip: bool,
-    app: bool,
-    triggerHair: bool,
-    triggerClick: bool,
-    trigger: f32,
-    touchpadTouch: bool,
-    touchpadClick: bool,
-    touchpad: math.Vector2f,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: ViveControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.grip, self.app, self.triggerHair, self.triggerClick, self.touchpadTouch, self.touchpadClick, false, false);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-        try ipc.write(@TypeOf(self.touchpad), self.touchpad);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ViveControllerState {
-        var self: ViveControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.grip, self.app, self.triggerHair, self.triggerClick, self.touchpadTouch, self.touchpadClick, _, _ = try ipc.read8PackedBools();
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        self.touchpad = try ipc.read(@TypeOf(self.touchpad));
-        return self;
-    }
-};
-
-pub const WindowsMR_ControllerState = struct {
-    grip: bool,
-    app: bool,
-    triggerHair: bool,
-    triggerClick: bool,
-    trigger: f32,
-    touchpadTouch: bool,
-    touchpadClick: bool,
-    touchpad: math.Vector2f,
-    joystickClick: bool,
-    joystickRaw: math.Vector2f,
-    deviceID: []const u16,
-    deviceModel: []const u16,
-    side: Chirality,
-    bodyNode: BodyNode,
-    isDeviceActive: bool,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    hasBoundHand: bool,
-    handPosition: math.Vector3f,
-    handRotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: WindowsMR_ControllerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.deviceID), self.deviceID);
-        try ipc.write(@TypeOf(self.deviceModel), self.deviceModel);
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.bodyNode), self.bodyNode);
-        try ipc.write8PackedBools(self.isDeviceActive, self.isTracking, self.hasBoundHand, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0053: brfalse.s IL_00a5
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0073: brfalse.s IL_008d
-        try ipc.write(@TypeOf(self.handPosition), self.handPosition);
-        try ipc.write(@TypeOf(self.handRotation), self.handRotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.batteryCharging), self.batteryCharging);
-        try ipc.write8PackedBools(self.grip, self.app, self.triggerHair, self.triggerClick, self.touchpadTouch, self.touchpadClick, self.joystickClick, false);
-        try ipc.write(@TypeOf(self.trigger), self.trigger);
-        try ipc.write(@TypeOf(self.touchpad), self.touchpad);
-        try ipc.write(@TypeOf(self.joystickRaw), self.joystickRaw);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !WindowsMR_ControllerState {
-        var self: WindowsMR_ControllerState = undefined;
-        self.deviceID = try ipc.read(@TypeOf(self.deviceID));
-        self.deviceModel = try ipc.read(@TypeOf(self.deviceModel));
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.bodyNode = try ipc.read(@TypeOf(self.bodyNode));
-        self.isDeviceActive, self.isTracking, self.hasBoundHand, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_004e: brfalse.s IL_00a0
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_006e: brfalse.s IL_0088
-        self.handPosition = try ipc.read(@TypeOf(self.handPosition));
-        self.handRotation = try ipc.read(@TypeOf(self.handRotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.batteryCharging = try ipc.read(@TypeOf(self.batteryCharging));
-        self.grip, self.app, self.triggerHair, self.triggerClick, self.touchpadTouch, self.touchpadClick, self.joystickClick, _ = try ipc.read8PackedBools();
-        self.trigger = try ipc.read(@TypeOf(self.trigger));
-        self.touchpad = try ipc.read(@TypeOf(self.touchpad));
-        self.joystickRaw = try ipc.read(@TypeOf(self.joystickRaw));
-        return self;
-    }
-};
-
-pub const HeadsetState = struct {
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-    connectionType: HeadsetConnection,
-    headsetManufacturer: []const u16,
-    headsetModel: []const u16,
-
-    pub fn write(self: HeadsetState, ipc: IpcSerializer) !void {
-        try ipc.write8PackedBools(self.isTracking, self.batteryCharging, false, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_001e: brfalse.s IL_0044
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-        try ipc.write(@TypeOf(self.connectionType), self.connectionType);
-        try ipc.write(@TypeOf(self.headsetManufacturer), self.headsetManufacturer);
-        try ipc.write(@TypeOf(self.headsetModel), self.headsetModel);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !HeadsetState {
-        var self: HeadsetState = undefined;
-        self.isTracking, self.batteryCharging, _, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0018: brfalse.s IL_003e
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        self.connectionType = try ipc.read(@TypeOf(self.connectionType));
-        self.headsetManufacturer = try ipc.read(@TypeOf(self.headsetManufacturer));
-        self.headsetModel = try ipc.read(@TypeOf(self.headsetModel));
-        return self;
-    }
-};
-
-pub const TrackerState = struct {
-    uniqueId: []const u16,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-    batteryLevel: f32,
-    batteryCharging: bool,
-
-    pub fn write(self: TrackerState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.uniqueId), self.uniqueId);
-        try ipc.write8PackedBools(self.isTracking, self.batteryCharging, false, false, false, false, false, false);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_002a: brfalse.s IL_0044
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-        try ipc.write(@TypeOf(self.batteryLevel), self.batteryLevel);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrackerState {
-        var self: TrackerState = undefined;
-        self.uniqueId = try ipc.read(@TypeOf(self.uniqueId));
-        self.isTracking, self.batteryCharging, _, _, _, _, _, _ = try ipc.read8PackedBools();
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_0024: brfalse.s IL_003e
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        self.batteryLevel = try ipc.read(@TypeOf(self.batteryLevel));
-        return self;
-    }
-};
-
-pub const TrackingReferenceState = struct {
-    uniqueId: []const u16,
-    isTracking: bool,
-    position: math.Vector3f,
-    rotation: math.Quaternionf,
-
-    pub fn write(self: TrackingReferenceState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.uniqueId), self.uniqueId);
-        try ipc.write(@TypeOf(self.isTracking), self.isTracking);
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_001e: brfalse.s IL_0038
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrackingReferenceState {
-        var self: TrackingReferenceState = undefined;
-        self.uniqueId = try ipc.read(@TypeOf(self.uniqueId));
-        self.isTracking = try ipc.read(@TypeOf(self.isTracking));
-        // FIXME: Unknown Cond_Branch instruction:
-        // IL_001e: brfalse.s IL_0038
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        return self;
-    }
-};
-
-pub const VR_InputsState = struct {
-    userPresentInHeadset: bool,
-    dashboardOpen: bool,
-    headsetState: ?HeadsetState,
-    controllers: []const VR_ControllerState,
-    trackers: []const TrackerState,
-    trackingReferences: []const TrackingReferenceState,
-    hands: []const HandState,
-    viveHandTracking: ?ViveHandTrackingInputState,
-
-    pub fn write(self: VR_InputsState, ipc: IpcSerializer) !void {
-        try ipc.write8PackedBools(self.userPresentInHeadset, self.dashboardOpen, false, false, false, false, false, false);
-        try ipc.write(@TypeOf(self.headsetState), self.headsetState);
-        try ipc.writePolymorphicList(@TypeOf(self.controllers), self.controllers);
-        try ipc.writeList(@TypeOf(self.trackers), self.trackers);
-        try ipc.writeList(@TypeOf(self.trackingReferences), self.trackingReferences);
-        try ipc.writeList(@TypeOf(self.hands), self.hands);
-        try ipc.write(@TypeOf(self.viveHandTracking), self.viveHandTracking);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VR_InputsState {
-        var self: VR_InputsState = undefined;
-        self.userPresentInHeadset, self.dashboardOpen, _, _, _, _, _, _ = try ipc.read8PackedBools();
-        self.headsetState = try ipc.read(@TypeOf(self.headsetState));
-        self.controllers = try ipc.readPolymorphicList(@TypeOf(self.controllers));
-        self.trackers = try ipc.readList(@TypeOf(self.trackers));
-        self.trackingReferences = try ipc.readList(@TypeOf(self.trackingReferences));
-        self.hands = try ipc.readList(@TypeOf(self.hands));
-        self.viveHandTracking = try ipc.read(@TypeOf(self.viveHandTracking));
-        return self;
-    }
-};
-
-pub const WindowState = struct {
-    isWindowFocused: bool,
-    isFullscreen: bool,
-    windowResolution: math.Vector2i,
-    resolutionSettingsApplied: bool,
-    dragAndDropEvent: ?DragAndDropEvent,
-
-    pub fn write(self: WindowState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.isWindowFocused), self.isWindowFocused);
-        try ipc.write(@TypeOf(self.isFullscreen), self.isFullscreen);
-        try ipc.write(@TypeOf(self.windowResolution), self.windowResolution);
-        try ipc.write(@TypeOf(self.resolutionSettingsApplied), self.resolutionSettingsApplied);
-        try ipc.write(@TypeOf(self.dragAndDropEvent), self.dragAndDropEvent);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !WindowState {
-        var self: WindowState = undefined;
-        self.isWindowFocused = try ipc.read(@TypeOf(self.isWindowFocused));
-        self.isFullscreen = try ipc.read(@TypeOf(self.isFullscreen));
-        self.windowResolution = try ipc.read(@TypeOf(self.windowResolution));
-        self.resolutionSettingsApplied = try ipc.read(@TypeOf(self.resolutionSettingsApplied));
-        self.dragAndDropEvent = try ipc.read(@TypeOf(self.dragAndDropEvent));
-        return self;
-    }
-};
-
-pub const OutputState = struct {
-    lockCursor: bool,
-    lockCursorPosition: ?math.Vector2i,
-    keyboardInputActive: bool,
-    vr: ?VR_OutputState,
-
-    pub fn write(self: OutputState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.lockCursor), self.lockCursor);
-        try ipc.write(@TypeOf(self.lockCursorPosition), self.lockCursorPosition);
-        try ipc.write(@TypeOf(self.keyboardInputActive), self.keyboardInputActive);
-        try ipc.write(@TypeOf(self.vr), self.vr);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !OutputState {
-        var self: OutputState = undefined;
-        self.lockCursor = try ipc.read(@TypeOf(self.lockCursor));
-        self.lockCursorPosition = try ipc.read(@TypeOf(self.lockCursorPosition));
-        self.keyboardInputActive = try ipc.read(@TypeOf(self.keyboardInputActive));
-        self.vr = try ipc.read(@TypeOf(self.vr));
-        return self;
-    }
-};
-
-pub const VR_ControllerOutputState = struct {
-    side: Chirality,
-    vibrateTime: f64,
-    hapticState: HapticPointState,
-
-    pub fn write(self: VR_ControllerOutputState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.side), self.side);
-        try ipc.write(@TypeOf(self.vibrateTime), self.vibrateTime);
-        try ipc.write(@TypeOf(self.hapticState), self.hapticState);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VR_ControllerOutputState {
-        var self: VR_ControllerOutputState = undefined;
-        self.side = try ipc.read(@TypeOf(self.side));
-        self.vibrateTime = try ipc.read(@TypeOf(self.vibrateTime));
-        self.hapticState = try ipc.read(@TypeOf(self.hapticState));
-        return self;
-    }
-};
-
-pub const VR_OutputState = struct {
-    leftController: ?VR_ControllerOutputState,
-    rightController: ?VR_ControllerOutputState,
-    useViveHandTracking: bool,
-
-    pub fn write(self: VR_OutputState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.leftController), self.leftController);
-        try ipc.write(@TypeOf(self.rightController), self.rightController);
-        try ipc.write(@TypeOf(self.useViveHandTracking), self.useViveHandTracking);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VR_OutputState {
-        var self: VR_OutputState = undefined;
-        self.leftController = try ipc.read(@TypeOf(self.leftController));
-        self.rightController = try ipc.read(@TypeOf(self.rightController));
-        self.useViveHandTracking = try ipc.read(@TypeOf(self.useViveHandTracking));
-        return self;
-    }
-};
-
-pub const BillboardRenderBufferUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: BillboardRenderBufferUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BillboardRenderBufferUpdate {
-        var self: BillboardRenderBufferUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const BlitToDisplayRenderablesUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: BlitToDisplayRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BlitToDisplayRenderablesUpdate {
-        var self: BlitToDisplayRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const CameraPortalsRenderablesUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: CameraPortalsRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraPortalsRenderablesUpdate {
-        var self: CameraPortalsRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const CameraRenderablesUpdate = struct {
-    transformIds: SharedMemoryBufferDescriptor,
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: CameraRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-        try ipc.write(@TypeOf(self.transformIds), self.transformIds);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraRenderablesUpdate {
-        var self: CameraRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        self.transformIds = try ipc.read(@TypeOf(self.transformIds));
-        return self;
-    }
-};
-
-pub const RenderMaterialOverridesUpdate = struct {
-    materialOverrideStates: SharedMemoryBufferDescriptor,
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: RenderMaterialOverridesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-        try ipc.write(@TypeOf(self.materialOverrideStates), self.materialOverrideStates);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderMaterialOverridesUpdate {
-        var self: RenderMaterialOverridesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        self.materialOverrideStates = try ipc.read(@TypeOf(self.materialOverrideStates));
-        return self;
-    }
-};
-
-pub const RenderTransformOverridesUpdate = struct {
-    skinnedMeshRenderersIndexes: SharedMemoryBufferDescriptor,
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: RenderTransformOverridesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-        try ipc.write(@TypeOf(self.skinnedMeshRenderersIndexes), self.skinnedMeshRenderersIndexes);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderTransformOverridesUpdate {
-        var self: RenderTransformOverridesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        self.skinnedMeshRenderersIndexes = try ipc.read(@TypeOf(self.skinnedMeshRenderersIndexes));
-        return self;
-    }
-};
-
-pub const GaussianSplatRenderablesUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: GaussianSplatRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !GaussianSplatRenderablesUpdate {
-        var self: GaussianSplatRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const LayerUpdate = struct {
-    layerAssignments: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: LayerUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.layerAssignments), self.layerAssignments);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LayerUpdate {
-        var self: LayerUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.layerAssignments = try ipc.read(@TypeOf(self.layerAssignments));
-        return self;
-    }
-};
-
-pub const LightsBufferRendererConsumed = struct {
-    globalUniqueId: i32,
-
-    pub fn write(self: LightsBufferRendererConsumed, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.globalUniqueId), self.globalUniqueId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightsBufferRendererConsumed {
-        var self: LightsBufferRendererConsumed = undefined;
-        self.globalUniqueId = try ipc.read(@TypeOf(self.globalUniqueId));
-        return self;
-    }
-};
-
-pub const LightsBufferRendererSubmission = struct {
-    lightsBufferUniqueId: i32,
-    lightsCount: i32,
-    lights: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: LightsBufferRendererSubmission, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.lightsBufferUniqueId), self.lightsBufferUniqueId);
-        try ipc.write(@TypeOf(self.lightsCount), self.lightsCount);
-        try ipc.write(@TypeOf(self.lights), self.lights);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightsBufferRendererSubmission {
-        var self: LightsBufferRendererSubmission = undefined;
-        self.lightsBufferUniqueId = try ipc.read(@TypeOf(self.lightsBufferUniqueId));
-        self.lightsCount = try ipc.read(@TypeOf(self.lightsCount));
-        self.lights = try ipc.read(@TypeOf(self.lights));
-        return self;
-    }
-};
-
-pub const LightsBufferRendererUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: LightsBufferRendererUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightsBufferRendererUpdate {
-        var self: LightsBufferRendererUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const LightRenderablesUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: LightRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightRenderablesUpdate {
-        var self: LightRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const LODGroupRenderablesUpdate = struct {
-    lodStates: SharedMemoryBufferDescriptor,
-    packedMeshRendererIds: SharedMemoryBufferDescriptor,
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: LODGroupRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-        try ipc.write(@TypeOf(self.lodStates), self.lodStates);
-        try ipc.write(@TypeOf(self.packedMeshRendererIds), self.packedMeshRendererIds);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LODGroupRenderablesUpdate {
-        var self: LODGroupRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        self.lodStates = try ipc.read(@TypeOf(self.lodStates));
-        self.packedMeshRendererIds = try ipc.read(@TypeOf(self.packedMeshRendererIds));
-        return self;
-    }
-};
-
-pub const MeshRenderBufferUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: MeshRenderBufferUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshRenderBufferUpdate {
-        var self: MeshRenderBufferUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const MeshRenderablesUpdate = struct {
-    meshStates: SharedMemoryBufferDescriptor,
-    meshMaterialsAndPropertyBlocks: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: MeshRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.meshStates), self.meshStates);
-        try ipc.write(@TypeOf(self.meshMaterialsAndPropertyBlocks), self.meshMaterialsAndPropertyBlocks);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshRenderablesUpdate {
-        var self: MeshRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.meshStates = try ipc.read(@TypeOf(self.meshStates));
-        self.meshMaterialsAndPropertyBlocks = try ipc.read(@TypeOf(self.meshMaterialsAndPropertyBlocks));
-        return self;
-    }
-};
-
-pub const ReflectionProbeRenderablesUpdate = struct {
-    changedProbesToRender: SharedMemoryBufferDescriptor,
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: ReflectionProbeRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-        try ipc.write(@TypeOf(self.changedProbesToRender), self.changedProbesToRender);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ReflectionProbeRenderablesUpdate {
-        var self: ReflectionProbeRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        self.changedProbesToRender = try ipc.read(@TypeOf(self.changedProbesToRender));
-        return self;
-    }
-};
-
-pub const SkinnedMeshRenderablesUpdate = struct {
-    boundsUpdates: SharedMemoryBufferDescriptor,
-    realtimeBoundsUpdates: SharedMemoryBufferDescriptor,
-    boneAssignments: SharedMemoryBufferDescriptor,
-    boneTransformIndexes: SharedMemoryBufferDescriptor,
-    blendshapeUpdateBatches: SharedMemoryBufferDescriptor,
-    blendshapeUpdates: SharedMemoryBufferDescriptor,
-    meshStates: SharedMemoryBufferDescriptor,
-    meshMaterialsAndPropertyBlocks: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: SkinnedMeshRenderablesUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.meshStates), self.meshStates);
-        try ipc.write(@TypeOf(self.meshMaterialsAndPropertyBlocks), self.meshMaterialsAndPropertyBlocks);
-        try ipc.write(@TypeOf(self.boundsUpdates), self.boundsUpdates);
-        try ipc.write(@TypeOf(self.realtimeBoundsUpdates), self.realtimeBoundsUpdates);
-        try ipc.write(@TypeOf(self.boneAssignments), self.boneAssignments);
-        try ipc.write(@TypeOf(self.boneTransformIndexes), self.boneTransformIndexes);
-        try ipc.write(@TypeOf(self.blendshapeUpdateBatches), self.blendshapeUpdateBatches);
-        try ipc.write(@TypeOf(self.blendshapeUpdates), self.blendshapeUpdates);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SkinnedMeshRenderablesUpdate {
-        var self: SkinnedMeshRenderablesUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.meshStates = try ipc.read(@TypeOf(self.meshStates));
-        self.meshMaterialsAndPropertyBlocks = try ipc.read(@TypeOf(self.meshMaterialsAndPropertyBlocks));
-        self.boundsUpdates = try ipc.read(@TypeOf(self.boundsUpdates));
-        self.realtimeBoundsUpdates = try ipc.read(@TypeOf(self.realtimeBoundsUpdates));
-        self.boneAssignments = try ipc.read(@TypeOf(self.boneAssignments));
-        self.boneTransformIndexes = try ipc.read(@TypeOf(self.boneTransformIndexes));
-        self.blendshapeUpdateBatches = try ipc.read(@TypeOf(self.blendshapeUpdateBatches));
-        self.blendshapeUpdates = try ipc.read(@TypeOf(self.blendshapeUpdates));
-        return self;
-    }
-};
-
-pub const ReflectionProbeSH2Tasks = struct {
-    tasks: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: ReflectionProbeSH2Tasks, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.tasks), self.tasks);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ReflectionProbeSH2Tasks {
-        var self: ReflectionProbeSH2Tasks = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.tasks = try ipc.read(@TypeOf(self.tasks));
-        return self;
-    }
-};
-
-pub const TrailsRendererUpdate = struct {
-    states: SharedMemoryBufferDescriptor,
-    removals: SharedMemoryBufferDescriptor,
-    additions: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: TrailsRendererUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.additions), self.additions);
-        try ipc.write(@TypeOf(self.states), self.states);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrailsRendererUpdate {
-        var self: TrailsRendererUpdate = undefined;
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.additions = try ipc.read(@TypeOf(self.additions));
-        self.states = try ipc.read(@TypeOf(self.states));
-        return self;
-    }
-};
-
-pub const TransformsUpdate = struct {
-    targetTransformCount: i32,
-    removals: SharedMemoryBufferDescriptor,
-    parentUpdates: SharedMemoryBufferDescriptor,
-    poseUpdates: SharedMemoryBufferDescriptor,
-
-    pub fn write(self: TransformsUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.targetTransformCount), self.targetTransformCount);
-        try ipc.write(@TypeOf(self.removals), self.removals);
-        try ipc.write(@TypeOf(self.parentUpdates), self.parentUpdates);
-        try ipc.write(@TypeOf(self.poseUpdates), self.poseUpdates);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TransformsUpdate {
-        var self: TransformsUpdate = undefined;
-        self.targetTransformCount = try ipc.read(@TypeOf(self.targetTransformCount));
-        self.removals = try ipc.read(@TypeOf(self.removals));
-        self.parentUpdates = try ipc.read(@TypeOf(self.parentUpdates));
-        self.poseUpdates = try ipc.read(@TypeOf(self.poseUpdates));
-        return self;
-    }
-};
-
-pub const RendererInitData = struct {
-    sharedMemoryPrefix: []const u16,
-    mainProcessId: i32,
-    debugFramePacing: bool,
-    outputDevice: HeadOutputDevice,
-    windowTitle: []const u16,
-
-    pub fn write(self: RendererInitData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.sharedMemoryPrefix), self.sharedMemoryPrefix);
-        try ipc.write(@TypeOf(self.mainProcessId), self.mainProcessId);
-        try ipc.write(@TypeOf(self.debugFramePacing), self.debugFramePacing);
-        try ipc.write(@TypeOf(self.outputDevice), self.outputDevice);
-        try ipc.write(@TypeOf(self.windowTitle), self.windowTitle);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RendererInitData {
-        var self: RendererInitData = undefined;
-        self.sharedMemoryPrefix = try ipc.read(@TypeOf(self.sharedMemoryPrefix));
-        self.mainProcessId = try ipc.read(@TypeOf(self.mainProcessId));
-        self.debugFramePacing = try ipc.read(@TypeOf(self.debugFramePacing));
-        self.outputDevice = try ipc.read(@TypeOf(self.outputDevice));
-        self.windowTitle = try ipc.read(@TypeOf(self.windowTitle));
-        return self;
-    }
-};
-
-pub const RendererInitFinalizeData = struct {};
-
-pub const RendererInitProgressUpdate = struct {
-    phaseIndex: i32,
-    phase: []const u16,
-    subPhase: []const u16,
-    forceShow: bool,
-
-    pub fn write(self: RendererInitProgressUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.phaseIndex), self.phaseIndex);
-        try ipc.write(@TypeOf(self.phase), self.phase);
-        try ipc.write(@TypeOf(self.subPhase), self.subPhase);
-        try ipc.write(@TypeOf(self.forceShow), self.forceShow);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RendererInitProgressUpdate {
-        var self: RendererInitProgressUpdate = undefined;
-        self.phaseIndex = try ipc.read(@TypeOf(self.phaseIndex));
-        self.phase = try ipc.read(@TypeOf(self.phase));
-        self.subPhase = try ipc.read(@TypeOf(self.subPhase));
-        self.forceShow = try ipc.read(@TypeOf(self.forceShow));
-        return self;
-    }
-};
-
-pub const RendererInitResult = struct {
-    actualOutputDevice: HeadOutputDevice,
-    rendererIdentifier: []const u16,
-    mainWindowHandlePtr: i64,
-    stereoRenderingMode: []const u16,
-    maxTextureSize: i32,
-    isGPUTexturePOTByteAligned: bool,
-    supportedTextureFormats: []const TextureFormat,
-
-    pub fn write(self: RendererInitResult, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.actualOutputDevice), self.actualOutputDevice);
-        try ipc.write(@TypeOf(self.rendererIdentifier), self.rendererIdentifier);
-        try ipc.write(@TypeOf(self.mainWindowHandlePtr), self.mainWindowHandlePtr);
-        try ipc.write(@TypeOf(self.stereoRenderingMode), self.stereoRenderingMode);
-        try ipc.write(@TypeOf(self.maxTextureSize), self.maxTextureSize);
-        try ipc.write(@TypeOf(self.isGPUTexturePOTByteAligned), self.isGPUTexturePOTByteAligned);
-        try ipc.writeList(@TypeOf(self.supportedTextureFormats), self.supportedTextureFormats);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RendererInitResult {
-        var self: RendererInitResult = undefined;
-        self.actualOutputDevice = try ipc.read(@TypeOf(self.actualOutputDevice));
-        self.rendererIdentifier = try ipc.read(@TypeOf(self.rendererIdentifier));
-        self.mainWindowHandlePtr = try ipc.read(@TypeOf(self.mainWindowHandlePtr));
-        self.stereoRenderingMode = try ipc.read(@TypeOf(self.stereoRenderingMode));
-        self.maxTextureSize = try ipc.read(@TypeOf(self.maxTextureSize));
-        self.isGPUTexturePOTByteAligned = try ipc.read(@TypeOf(self.isGPUTexturePOTByteAligned));
-        self.supportedTextureFormats = try ipc.readList(@TypeOf(self.supportedTextureFormats));
-        return self;
-    }
-};
-
-pub const RendererParentWindow = struct {
-    windowHandle: i64,
-
-    pub fn write(self: RendererParentWindow, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.windowHandle), self.windowHandle);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RendererParentWindow {
-        var self: RendererParentWindow = undefined;
-        self.windowHandle = try ipc.read(@TypeOf(self.windowHandle));
-        return self;
-    }
-};
-
-pub const RendererShutdown = struct {};
-
-pub const RendererShutdownRequest = struct {};
-
-pub const RenderBoundingBox = extern struct {
-    center: math.Vector3f,
-    extents: math.Vector3f,
-
-    pub fn write(self: RenderBoundingBox, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.center), self.center);
-        try ipc.write(@TypeOf(self.extents), self.extents);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderBoundingBox {
-        var self: RenderBoundingBox = undefined;
-        self.center = try ipc.read(@TypeOf(self.center));
-        self.extents = try ipc.read(@TypeOf(self.extents));
-        return self;
-    }
-};
-
-pub const MaterialPropertyUpdate = extern struct {
-    propertyID: i32,
-    updateType: MaterialPropertyUpdateType,
-
-    pub fn write(self: MaterialPropertyUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.propertyID), self.propertyID);
-        try ipc.write(@TypeOf(self.updateType), self.updateType);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MaterialPropertyUpdate {
-        var self: MaterialPropertyUpdate = undefined;
-        self.propertyID = try ipc.read(@TypeOf(self.propertyID));
-        self.updateType = try ipc.read(@TypeOf(self.updateType));
-        return self;
-    }
-};
-
-pub const VertexAttributeDescriptor = extern struct {
-    attribute: VertexAttributeType,
-    format: VertexAttributeFormat,
-    dimensions: i32,
-
-    pub fn write(self: VertexAttributeDescriptor, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.attribute), self.attribute);
-        try ipc.write(@TypeOf(self.format), self.format);
-        try ipc.write(@TypeOf(self.dimensions), self.dimensions);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !VertexAttributeDescriptor {
-        var self: VertexAttributeDescriptor = undefined;
-        self.attribute = try ipc.read(@TypeOf(self.attribute));
-        self.format = try ipc.read(@TypeOf(self.format));
-        self.dimensions = try ipc.read(@TypeOf(self.dimensions));
-        return self;
-    }
-};
-
-pub const SubmeshBufferDescriptor = extern struct {
-    topology: SubmeshTopology,
-    indexStart: i32,
-    indexCount: i32,
-    bounds: RenderBoundingBox,
-
-    pub fn write(self: SubmeshBufferDescriptor, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.topology), self.topology);
-        try ipc.write(@TypeOf(self.indexStart), self.indexStart);
-        try ipc.write(@TypeOf(self.indexCount), self.indexCount);
-        try ipc.write(@TypeOf(self.bounds), self.bounds);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !SubmeshBufferDescriptor {
-        var self: SubmeshBufferDescriptor = undefined;
-        self.topology = try ipc.read(@TypeOf(self.topology));
-        self.indexStart = try ipc.read(@TypeOf(self.indexStart));
-        self.indexCount = try ipc.read(@TypeOf(self.indexCount));
-        self.bounds = try ipc.read(@TypeOf(self.bounds));
-        return self;
-    }
-};
-
-pub const BlendshapeBufferDescriptor = extern struct {
-    blendshapeIndex: i32,
-    frameIndex: i32,
-    frameWeight: f32,
-    dataFlags: BlendshapeDataFlags,
-
-    pub fn write(self: BlendshapeBufferDescriptor, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.blendshapeIndex), self.blendshapeIndex);
-        try ipc.write(@TypeOf(self.frameIndex), self.frameIndex);
-        try ipc.write(@TypeOf(self.frameWeight), self.frameWeight);
-        try ipc.write(@TypeOf(self.dataFlags), self.dataFlags);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BlendshapeBufferDescriptor {
-        var self: BlendshapeBufferDescriptor = undefined;
-        self.blendshapeIndex = try ipc.read(@TypeOf(self.blendshapeIndex));
-        self.frameIndex = try ipc.read(@TypeOf(self.frameIndex));
-        self.frameWeight = try ipc.read(@TypeOf(self.frameWeight));
-        self.dataFlags = try ipc.read(@TypeOf(self.dataFlags));
-        return self;
-    }
-};
-
-pub const MeshUploadHint = extern struct {};
-
-pub const TextureUploadHint = extern struct {
-    regionData: RenderIntRect,
-    readable: bool,
-    hasRegion: bool,
-
-    pub fn write(self: TextureUploadHint, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.regionData), self.regionData);
-        try ipc.write(@TypeOf(self.readable), self.readable);
-        // NOTE: field with gap/overlap, hasRegion = offset:17, size:4, gap:-3
-        try ipc.write(@TypeOf(self.hasRegion), self.hasRegion);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TextureUploadHint {
-        var self: TextureUploadHint = undefined;
-        self.regionData = try ipc.read(@TypeOf(self.regionData));
-        self.readable = try ipc.read(@TypeOf(self.readable));
-        // NOTE: field with gap/overlap, hasRegion = offset:17, size:4, gap:-3
-        self.hasRegion = try ipc.read(@TypeOf(self.hasRegion));
-        return self;
-    }
-};
-
-pub const Texture3DUploadHint = extern struct {
-    readable: bool,
-
-    pub fn write(self: Texture3DUploadHint, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.readable), self.readable);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !Texture3DUploadHint {
-        var self: Texture3DUploadHint = undefined;
-        self.readable = try ipc.read(@TypeOf(self.readable));
-        return self;
-    }
-};
-
-pub const RenderableHandle = extern struct {
-    renderSpaceId: i32,
-    renderableIndex: i32,
-
-    pub fn write(self: RenderableHandle, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderSpaceId), self.renderSpaceId);
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderableHandle {
-        var self: RenderableHandle = undefined;
-        self.renderSpaceId = try ipc.read(@TypeOf(self.renderSpaceId));
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        return self;
-    }
-};
-
-pub const RenderTransform = extern struct {
-    position: math.Vector3f,
-    scale: math.Vector3f,
-    rotation: math.Quaternionf,
-
-    pub fn write(self: RenderTransform, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.position), self.position);
-        try ipc.write(@TypeOf(self.scale), self.scale);
-        try ipc.write(@TypeOf(self.rotation), self.rotation);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderTransform {
-        var self: RenderTransform = undefined;
-        self.position = try ipc.read(@TypeOf(self.position));
-        self.scale = try ipc.read(@TypeOf(self.scale));
-        self.rotation = try ipc.read(@TypeOf(self.rotation));
-        return self;
-    }
-};
-
-pub const RenderSH2 = extern struct {
-    sh0: math.Vector3f,
-    sh1: math.Vector3f,
-    sh2: math.Vector3f,
-    sh3: math.Vector3f,
-    sh4: math.Vector3f,
-    sh5: math.Vector3f,
-    sh6: math.Vector3f,
-    sh7: math.Vector3f,
-    sh8: math.Vector3f,
-
-    pub fn write(self: RenderSH2, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.sh0), self.sh0);
-        try ipc.write(@TypeOf(self.sh1), self.sh1);
-        try ipc.write(@TypeOf(self.sh2), self.sh2);
-        try ipc.write(@TypeOf(self.sh3), self.sh3);
-        try ipc.write(@TypeOf(self.sh4), self.sh4);
-        try ipc.write(@TypeOf(self.sh5), self.sh5);
-        try ipc.write(@TypeOf(self.sh6), self.sh6);
-        try ipc.write(@TypeOf(self.sh7), self.sh7);
-        try ipc.write(@TypeOf(self.sh8), self.sh8);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderSH2 {
-        var self: RenderSH2 = undefined;
-        self.sh0 = try ipc.read(@TypeOf(self.sh0));
-        self.sh1 = try ipc.read(@TypeOf(self.sh1));
-        self.sh2 = try ipc.read(@TypeOf(self.sh2));
-        self.sh3 = try ipc.read(@TypeOf(self.sh3));
-        self.sh4 = try ipc.read(@TypeOf(self.sh4));
-        self.sh5 = try ipc.read(@TypeOf(self.sh5));
-        self.sh6 = try ipc.read(@TypeOf(self.sh6));
-        self.sh7 = try ipc.read(@TypeOf(self.sh7));
-        self.sh8 = try ipc.read(@TypeOf(self.sh8));
-        return self;
-    }
-};
-
-pub const HapticPointState = extern struct {
-    force: f32,
-    temperature: f32,
-    pain: f32,
-    vibration: f32,
-
-    pub fn write(self: HapticPointState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.force), self.force);
-        try ipc.write(@TypeOf(self.temperature), self.temperature);
-        try ipc.write(@TypeOf(self.pain), self.pain);
-        try ipc.write(@TypeOf(self.vibration), self.vibration);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !HapticPointState {
-        var self: HapticPointState = undefined;
-        self.force = try ipc.read(@TypeOf(self.force));
-        self.temperature = try ipc.read(@TypeOf(self.temperature));
-        self.pain = try ipc.read(@TypeOf(self.pain));
-        self.vibration = try ipc.read(@TypeOf(self.vibration));
-        return self;
-    }
-};
-
-pub const BillboardRenderBufferState = extern struct {
-    renderableIndex: i32,
-    pointRenderBufferAssetId: i32,
-    materialAssetId: i32,
-    minBillboardScreenSize: f32,
-    maxBillboardScreenSize: f32,
-    alignment: BillboardAlignment,
-    motionVectorMode: MotionVectorMode,
-
-    pub fn write(self: BillboardRenderBufferState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.pointRenderBufferAssetId), self.pointRenderBufferAssetId);
-        try ipc.write(@TypeOf(self.materialAssetId), self.materialAssetId);
-        try ipc.write(@TypeOf(self.minBillboardScreenSize), self.minBillboardScreenSize);
-        try ipc.write(@TypeOf(self.maxBillboardScreenSize), self.maxBillboardScreenSize);
-        try ipc.write(@TypeOf(self.alignment), self.alignment);
-        try ipc.write(@TypeOf(self.motionVectorMode), self.motionVectorMode);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BillboardRenderBufferState {
-        var self: BillboardRenderBufferState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.pointRenderBufferAssetId = try ipc.read(@TypeOf(self.pointRenderBufferAssetId));
-        self.materialAssetId = try ipc.read(@TypeOf(self.materialAssetId));
-        self.minBillboardScreenSize = try ipc.read(@TypeOf(self.minBillboardScreenSize));
-        self.maxBillboardScreenSize = try ipc.read(@TypeOf(self.maxBillboardScreenSize));
-        self.alignment = try ipc.read(@TypeOf(self.alignment));
-        self.motionVectorMode = try ipc.read(@TypeOf(self.motionVectorMode));
-        return self;
-    }
-};
-
-pub const BlitToDisplayState = extern struct {
-    renderableIndex: i32,
-    textureId: i32,
-    backgroundColor: math.Vector4f,
-    displayIndex: i16,
-
-    pub fn write(self: BlitToDisplayState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.textureId), self.textureId);
-        try ipc.write(@TypeOf(self.backgroundColor), self.backgroundColor);
-        try ipc.write(@TypeOf(self.displayIndex), self.displayIndex);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BlitToDisplayState {
-        var self: BlitToDisplayState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.textureId = try ipc.read(@TypeOf(self.textureId));
-        self.backgroundColor = try ipc.read(@TypeOf(self.backgroundColor));
-        self.displayIndex = try ipc.read(@TypeOf(self.displayIndex));
-        return self;
-    }
-};
-
-pub const CameraPortalState = extern struct {
-    renderableIndex: i32,
-    meshRendererIndex: i32,
-    planeNormal: math.Vector3f,
-    planeOffset: f32,
-    renderTextureId: i32,
-    portalTransform: math.Matrix4x4f,
-    portalPlanePosition: math.Vector3f,
-    portalPlaneNormal: math.Vector3f,
-
-    pub fn write(self: CameraPortalState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.meshRendererIndex), self.meshRendererIndex);
-        try ipc.write(@TypeOf(self.planeNormal), self.planeNormal);
-        try ipc.write(@TypeOf(self.planeOffset), self.planeOffset);
-        try ipc.write(@TypeOf(self.renderTextureId), self.renderTextureId);
-        try ipc.write(@TypeOf(self.portalTransform), self.portalTransform);
-        try ipc.write(@TypeOf(self.portalPlanePosition), self.portalPlanePosition);
-        try ipc.write(@TypeOf(self.portalPlaneNormal), self.portalPlaneNormal);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraPortalState {
-        var self: CameraPortalState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.meshRendererIndex = try ipc.read(@TypeOf(self.meshRendererIndex));
-        self.planeNormal = try ipc.read(@TypeOf(self.planeNormal));
-        self.planeOffset = try ipc.read(@TypeOf(self.planeOffset));
-        self.renderTextureId = try ipc.read(@TypeOf(self.renderTextureId));
-        self.portalTransform = try ipc.read(@TypeOf(self.portalTransform));
-        self.portalPlanePosition = try ipc.read(@TypeOf(self.portalPlanePosition));
-        self.portalPlaneNormal = try ipc.read(@TypeOf(self.portalPlaneNormal));
-        return self;
-    }
-};
-
-pub const CameraState = extern struct {
-    renderableIndex: i32,
-    fieldOfView: f32,
-    orthographicSize: f32,
-    nearClip: f32,
-    farClip: f32,
-    backgroundColor: math.Vector4f,
-    viewport: RenderRect,
-    depth: f32,
-    renderTextureAssetId: i32,
-    selectiveRenderCount: i32,
-    excludeRenderCount: i32,
-    clearMode: CameraClearMode,
-    projection: CameraProjection,
-
-    pub fn write(self: CameraState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.fieldOfView), self.fieldOfView);
-        try ipc.write(@TypeOf(self.orthographicSize), self.orthographicSize);
-        try ipc.write(@TypeOf(self.nearClip), self.nearClip);
-        try ipc.write(@TypeOf(self.farClip), self.farClip);
-        try ipc.write(@TypeOf(self.backgroundColor), self.backgroundColor);
-        try ipc.write(@TypeOf(self.viewport), self.viewport);
-        try ipc.write(@TypeOf(self.depth), self.depth);
-        try ipc.write(@TypeOf(self.renderTextureAssetId), self.renderTextureAssetId);
-        try ipc.write(@TypeOf(self.selectiveRenderCount), self.selectiveRenderCount);
-        try ipc.write(@TypeOf(self.excludeRenderCount), self.excludeRenderCount);
-        try ipc.write(@TypeOf(self.clearMode), self.clearMode);
-        try ipc.write(@TypeOf(self.projection), self.projection);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !CameraState {
-        var self: CameraState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.fieldOfView = try ipc.read(@TypeOf(self.fieldOfView));
-        self.orthographicSize = try ipc.read(@TypeOf(self.orthographicSize));
-        self.nearClip = try ipc.read(@TypeOf(self.nearClip));
-        self.farClip = try ipc.read(@TypeOf(self.farClip));
-        self.backgroundColor = try ipc.read(@TypeOf(self.backgroundColor));
-        self.viewport = try ipc.read(@TypeOf(self.viewport));
-        self.depth = try ipc.read(@TypeOf(self.depth));
-        self.renderTextureAssetId = try ipc.read(@TypeOf(self.renderTextureAssetId));
-        self.selectiveRenderCount = try ipc.read(@TypeOf(self.selectiveRenderCount));
-        self.excludeRenderCount = try ipc.read(@TypeOf(self.excludeRenderCount));
-        self.clearMode = try ipc.read(@TypeOf(self.clearMode));
-        self.projection = try ipc.read(@TypeOf(self.projection));
-        return self;
-    }
-};
-
-pub const MaterialOverrideState = struct {
-    materialSlotIndex: i32,
-    materialAssetId: i32,
-};
-
-pub const RenderMaterialOverrideState = extern struct {
-    renderableIndex: i32,
-    packedMeshRendererIndex: i32,
-    materrialOverrideCount: i16,
-    context: RenderingContext,
-
-    pub fn write(self: RenderMaterialOverrideState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.packedMeshRendererIndex), self.packedMeshRendererIndex);
-        try ipc.write(@TypeOf(self.materrialOverrideCount), self.materrialOverrideCount);
-        try ipc.write(@TypeOf(self.context), self.context);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderMaterialOverrideState {
-        var self: RenderMaterialOverrideState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.packedMeshRendererIndex = try ipc.read(@TypeOf(self.packedMeshRendererIndex));
-        self.materrialOverrideCount = try ipc.read(@TypeOf(self.materrialOverrideCount));
-        self.context = try ipc.read(@TypeOf(self.context));
-        return self;
-    }
-};
-
-pub const RenderTransformOverrideState = extern struct {
-    renderableIndex: i32,
-    positionOverride: math.Vector3f,
-    rotationOverride: math.Quaternionf,
-    scaleOverride: math.Vector3f,
-    skinnedMeshRendererCount: i32,
-    context: RenderingContext,
-    overrideFlags: u8,
-
-    pub fn write(self: RenderTransformOverrideState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.positionOverride), self.positionOverride);
-        // NOTE: field with gap/overlap, rotationOverride = offset:20, size:16, gap:4
-        try ipc.write(@TypeOf(self.rotationOverride), self.rotationOverride);
-        try ipc.write(@TypeOf(self.scaleOverride), self.scaleOverride);
-        try ipc.write(@TypeOf(self.skinnedMeshRendererCount), self.skinnedMeshRendererCount);
-        try ipc.write(@TypeOf(self.context), self.context);
-        try ipc.write(@TypeOf(self.overrideFlags), self.overrideFlags);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderTransformOverrideState {
-        var self: RenderTransformOverrideState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.positionOverride = try ipc.read(@TypeOf(self.positionOverride));
-        // NOTE: field with gap/overlap, rotationOverride = offset:20, size:16, gap:4
-        self.rotationOverride = try ipc.read(@TypeOf(self.rotationOverride));
-        self.scaleOverride = try ipc.read(@TypeOf(self.scaleOverride));
-        self.skinnedMeshRendererCount = try ipc.read(@TypeOf(self.skinnedMeshRendererCount));
-        self.context = try ipc.read(@TypeOf(self.context));
-        self.overrideFlags = try ipc.read(@TypeOf(self.overrideFlags));
-        return self;
-    }
-};
-
-pub const GaussianSplatRendererState = struct {
-    renderableIndex: i32,
-    gaussianSplatAssetId: i32,
-    sizeScale: f32,
-    opacityScale: f32,
-    maxSHOrder: i32,
-    sphericalHamornicsOnly: bool,
-};
-
-pub const LightData = extern struct {
-    point: math.Vector3f,
-    orientation: math.Quaternionf,
-    color: math.Vector3f,
-    intensity: f32,
-    range: f32,
-    angle: f32,
-
-    pub fn write(self: LightData, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.point), self.point);
-        try ipc.write(@TypeOf(self.orientation), self.orientation);
-        try ipc.write(@TypeOf(self.color), self.color);
-        try ipc.write(@TypeOf(self.intensity), self.intensity);
-        try ipc.write(@TypeOf(self.range), self.range);
-        try ipc.write(@TypeOf(self.angle), self.angle);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightData {
-        var self: LightData = undefined;
-        self.point = try ipc.read(@TypeOf(self.point));
-        self.orientation = try ipc.read(@TypeOf(self.orientation));
-        self.color = try ipc.read(@TypeOf(self.color));
-        self.intensity = try ipc.read(@TypeOf(self.intensity));
-        self.range = try ipc.read(@TypeOf(self.range));
-        self.angle = try ipc.read(@TypeOf(self.angle));
-        return self;
-    }
-};
-
-pub const LightsBufferRendererState = extern struct {
-    renderableIndex: i32,
-    globalUniqueId: i32,
-    shadowStrength: f32,
-    shadowNearPlane: f32,
-    shadowMapResolution: i32,
-    shadowBias: f32,
-    shadowNormalBias: f32,
-    cookieTextureAssetId: i32,
-    lightType: LightType,
-    shadowType: ShadowType,
-
-    pub fn write(self: LightsBufferRendererState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.globalUniqueId), self.globalUniqueId);
-        try ipc.write(@TypeOf(self.shadowStrength), self.shadowStrength);
-        try ipc.write(@TypeOf(self.shadowNearPlane), self.shadowNearPlane);
-        try ipc.write(@TypeOf(self.shadowMapResolution), self.shadowMapResolution);
-        try ipc.write(@TypeOf(self.shadowBias), self.shadowBias);
-        try ipc.write(@TypeOf(self.shadowNormalBias), self.shadowNormalBias);
-        try ipc.write(@TypeOf(self.cookieTextureAssetId), self.cookieTextureAssetId);
-        try ipc.write(@TypeOf(self.lightType), self.lightType);
-        try ipc.write(@TypeOf(self.shadowType), self.shadowType);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightsBufferRendererState {
-        var self: LightsBufferRendererState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.globalUniqueId = try ipc.read(@TypeOf(self.globalUniqueId));
-        self.shadowStrength = try ipc.read(@TypeOf(self.shadowStrength));
-        self.shadowNearPlane = try ipc.read(@TypeOf(self.shadowNearPlane));
-        self.shadowMapResolution = try ipc.read(@TypeOf(self.shadowMapResolution));
-        self.shadowBias = try ipc.read(@TypeOf(self.shadowBias));
-        self.shadowNormalBias = try ipc.read(@TypeOf(self.shadowNormalBias));
-        self.cookieTextureAssetId = try ipc.read(@TypeOf(self.cookieTextureAssetId));
-        self.lightType = try ipc.read(@TypeOf(self.lightType));
-        self.shadowType = try ipc.read(@TypeOf(self.shadowType));
-        return self;
-    }
-};
-
-pub const LightState = extern struct {
-    renderableIndex: i32,
-    intensity: f32,
-    range: f32,
-    spotAngle: f32,
-    color: math.Vector4f,
-    shadowStrength: f32,
-    shadowNearPlane: f32,
-    shadowMapResolutionOverride: i32,
-    shadowBias: f32,
-    shadowNormalBias: f32,
-    cookieTextureAssetId: i32,
-    type: LightType,
-    shadowType: ShadowType,
-
-    pub fn write(self: LightState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.intensity), self.intensity);
-        try ipc.write(@TypeOf(self.range), self.range);
-        try ipc.write(@TypeOf(self.spotAngle), self.spotAngle);
-        try ipc.write(@TypeOf(self.color), self.color);
-        try ipc.write(@TypeOf(self.shadowStrength), self.shadowStrength);
-        try ipc.write(@TypeOf(self.shadowNearPlane), self.shadowNearPlane);
-        try ipc.write(@TypeOf(self.shadowMapResolutionOverride), self.shadowMapResolutionOverride);
-        try ipc.write(@TypeOf(self.shadowBias), self.shadowBias);
-        try ipc.write(@TypeOf(self.shadowNormalBias), self.shadowNormalBias);
-        try ipc.write(@TypeOf(self.cookieTextureAssetId), self.cookieTextureAssetId);
-        try ipc.write(@TypeOf(self.type), self.type);
-        try ipc.write(@TypeOf(self.shadowType), self.shadowType);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !LightState {
-        var self: LightState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.intensity = try ipc.read(@TypeOf(self.intensity));
-        self.range = try ipc.read(@TypeOf(self.range));
-        self.spotAngle = try ipc.read(@TypeOf(self.spotAngle));
-        self.color = try ipc.read(@TypeOf(self.color));
-        self.shadowStrength = try ipc.read(@TypeOf(self.shadowStrength));
-        self.shadowNearPlane = try ipc.read(@TypeOf(self.shadowNearPlane));
-        self.shadowMapResolutionOverride = try ipc.read(@TypeOf(self.shadowMapResolutionOverride));
-        self.shadowBias = try ipc.read(@TypeOf(self.shadowBias));
-        self.shadowNormalBias = try ipc.read(@TypeOf(self.shadowNormalBias));
-        self.cookieTextureAssetId = try ipc.read(@TypeOf(self.cookieTextureAssetId));
-        self.type = try ipc.read(@TypeOf(self.type));
-        self.shadowType = try ipc.read(@TypeOf(self.shadowType));
-        return self;
-    }
-};
-
-pub const LODState = struct {
-    screenRelativeTransitionHeight: f32,
-    fadeTransitionWidth: f32,
-    rendererCount: i32,
-};
-
-pub const LODGroupState = struct {
-    renderableIndex: i32,
-    lodCount: i32,
-    crossFade: bool,
-    animateCrossFading: bool,
-};
-
-pub const MeshRenderBufferState = extern struct {
-    renderableIndex: i32,
-    pointRenderBufferAssetId: i32,
-    materialAssetId: i32,
-    meshAssetId: i32,
-    alignment: MeshAlignment,
-
-    pub fn write(self: MeshRenderBufferState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.pointRenderBufferAssetId), self.pointRenderBufferAssetId);
-        try ipc.write(@TypeOf(self.materialAssetId), self.materialAssetId);
-        try ipc.write(@TypeOf(self.meshAssetId), self.meshAssetId);
-        try ipc.write(@TypeOf(self.alignment), self.alignment);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshRenderBufferState {
-        var self: MeshRenderBufferState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.pointRenderBufferAssetId = try ipc.read(@TypeOf(self.pointRenderBufferAssetId));
-        self.materialAssetId = try ipc.read(@TypeOf(self.materialAssetId));
-        self.meshAssetId = try ipc.read(@TypeOf(self.meshAssetId));
-        self.alignment = try ipc.read(@TypeOf(self.alignment));
-        return self;
-    }
-};
-
-pub const MeshRendererState = extern struct {
-    renderableIndex: i32,
-    meshAssetId: i32,
-    materialCount: i32,
-    materialPropertyBlockCount: i32,
-    sortingOrder: i32,
-    shadowCastMode: ShadowCastMode,
-    motionVectorMode: MotionVectorMode,
-
-    pub fn write(self: MeshRendererState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.meshAssetId), self.meshAssetId);
-        try ipc.write(@TypeOf(self.materialCount), self.materialCount);
-        try ipc.write(@TypeOf(self.materialPropertyBlockCount), self.materialPropertyBlockCount);
-        try ipc.write(@TypeOf(self.sortingOrder), self.sortingOrder);
-        try ipc.write(@TypeOf(self.shadowCastMode), self.shadowCastMode);
-        try ipc.write(@TypeOf(self.motionVectorMode), self.motionVectorMode);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !MeshRendererState {
-        var self: MeshRendererState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.meshAssetId = try ipc.read(@TypeOf(self.meshAssetId));
-        self.materialCount = try ipc.read(@TypeOf(self.materialCount));
-        self.materialPropertyBlockCount = try ipc.read(@TypeOf(self.materialPropertyBlockCount));
-        self.sortingOrder = try ipc.read(@TypeOf(self.sortingOrder));
-        self.shadowCastMode = try ipc.read(@TypeOf(self.shadowCastMode));
-        self.motionVectorMode = try ipc.read(@TypeOf(self.motionVectorMode));
-        return self;
-    }
-};
-
-pub const ReflectionProbeState = extern struct {
-    renderableIndex: i32,
-    importance: i32,
-    intensity: f32,
-    blendDistance: f32,
-    boxSize: math.Vector3f,
-    cubemapAssetId: i32,
-    resolution: i32,
-    shadowDistance: f32,
-    backgroundColor: math.Vector4f,
-    nearClip: f32,
-    farClip: f32,
-    type: ReflectionProbeType,
-    clearFlags: ReflectionProbeClear,
-    timeSlicingMode: ReflectionProbeTimeSlicingMode,
-
-    pub fn write(self: ReflectionProbeState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.importance), self.importance);
-        try ipc.write(@TypeOf(self.intensity), self.intensity);
-        try ipc.write(@TypeOf(self.blendDistance), self.blendDistance);
-        try ipc.write(@TypeOf(self.boxSize), self.boxSize);
-        try ipc.write(@TypeOf(self.cubemapAssetId), self.cubemapAssetId);
-        try ipc.write(@TypeOf(self.resolution), self.resolution);
-        try ipc.write(@TypeOf(self.shadowDistance), self.shadowDistance);
-        try ipc.write(@TypeOf(self.backgroundColor), self.backgroundColor);
-        try ipc.write(@TypeOf(self.nearClip), self.nearClip);
-        try ipc.write(@TypeOf(self.farClip), self.farClip);
-        try ipc.write(@TypeOf(self.type), self.type);
-        try ipc.write(@TypeOf(self.clearFlags), self.clearFlags);
-        try ipc.write(@TypeOf(self.timeSlicingMode), self.timeSlicingMode);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !ReflectionProbeState {
-        var self: ReflectionProbeState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.importance = try ipc.read(@TypeOf(self.importance));
-        self.intensity = try ipc.read(@TypeOf(self.intensity));
-        self.blendDistance = try ipc.read(@TypeOf(self.blendDistance));
-        self.boxSize = try ipc.read(@TypeOf(self.boxSize));
-        self.cubemapAssetId = try ipc.read(@TypeOf(self.cubemapAssetId));
-        self.resolution = try ipc.read(@TypeOf(self.resolution));
-        self.shadowDistance = try ipc.read(@TypeOf(self.shadowDistance));
-        self.backgroundColor = try ipc.read(@TypeOf(self.backgroundColor));
-        self.nearClip = try ipc.read(@TypeOf(self.nearClip));
-        self.farClip = try ipc.read(@TypeOf(self.farClip));
-        self.type = try ipc.read(@TypeOf(self.type));
-        self.clearFlags = try ipc.read(@TypeOf(self.clearFlags));
-        self.timeSlicingMode = try ipc.read(@TypeOf(self.timeSlicingMode));
-        return self;
-    }
-};
-
-pub const SkinnedMeshBoundsUpdate = struct {
-    renderableIndex: i32,
-    localBounds: RenderBoundingBox,
-};
-
-pub const SkinnedMeshRealtimeBoundsUpdate = struct {
-    renderableIndex: i32,
-    computedGlobalBounds: RenderBoundingBox,
-};
-
-pub const BoneAssignment = extern struct {
-    renderableIndex: i32,
-    rootBoneTransformId: i32,
-    boneCount: i32,
-
-    pub fn write(self: BoneAssignment, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.rootBoneTransformId), self.rootBoneTransformId);
-        try ipc.write(@TypeOf(self.boneCount), self.boneCount);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BoneAssignment {
-        var self: BoneAssignment = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.rootBoneTransformId = try ipc.read(@TypeOf(self.rootBoneTransformId));
-        self.boneCount = try ipc.read(@TypeOf(self.boneCount));
-        return self;
-    }
-};
-
-pub const BlendshapeUpdateBatch = extern struct {
-    renderableIndex: i32,
-    blendshapeUpdateCount: i32,
-
-    pub fn write(self: BlendshapeUpdateBatch, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.blendshapeUpdateCount), self.blendshapeUpdateCount);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BlendshapeUpdateBatch {
-        var self: BlendshapeUpdateBatch = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.blendshapeUpdateCount = try ipc.read(@TypeOf(self.blendshapeUpdateCount));
-        return self;
-    }
-};
-
-pub const BlendshapeUpdate = extern struct {
-    blendshapeIndex: i32,
-    weight: f32,
-
-    pub fn write(self: BlendshapeUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.blendshapeIndex), self.blendshapeIndex);
-        try ipc.write(@TypeOf(self.weight), self.weight);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !BlendshapeUpdate {
-        var self: BlendshapeUpdate = undefined;
-        self.blendshapeIndex = try ipc.read(@TypeOf(self.blendshapeIndex));
-        self.weight = try ipc.read(@TypeOf(self.weight));
-        return self;
-    }
-};
-
-pub const ReflectionProbeSH2Task = struct {
-    renderableIndex: i32,
-    reflectionProbeRenderableIndex: i32,
-    result: ComputeResult,
-    resultData: RenderSH2,
-};
-
-pub const TrailsRendererState = extern struct {
-    renderableIndex: i32,
-    trailsRenderBufferAssetId: i32,
-    materialAssetId: i32,
-    textureMode: TrailTextureMode,
-    motionVectorMode: MotionVectorMode,
-    generateLightingData: bool,
-
-    pub fn write(self: TrailsRendererState, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.renderableIndex), self.renderableIndex);
-        try ipc.write(@TypeOf(self.trailsRenderBufferAssetId), self.trailsRenderBufferAssetId);
-        try ipc.write(@TypeOf(self.materialAssetId), self.materialAssetId);
-        try ipc.write(@TypeOf(self.textureMode), self.textureMode);
-        try ipc.write(@TypeOf(self.motionVectorMode), self.motionVectorMode);
-        try ipc.write(@TypeOf(self.generateLightingData), self.generateLightingData);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TrailsRendererState {
-        var self: TrailsRendererState = undefined;
-        self.renderableIndex = try ipc.read(@TypeOf(self.renderableIndex));
-        self.trailsRenderBufferAssetId = try ipc.read(@TypeOf(self.trailsRenderBufferAssetId));
-        self.materialAssetId = try ipc.read(@TypeOf(self.materialAssetId));
-        self.textureMode = try ipc.read(@TypeOf(self.textureMode));
-        self.motionVectorMode = try ipc.read(@TypeOf(self.motionVectorMode));
-        self.generateLightingData = try ipc.read(@TypeOf(self.generateLightingData));
-        return self;
-    }
-};
-
-pub const TransformParentUpdate = extern struct {
-    transformId: i32,
-    newParentId: i32,
-
-    pub fn write(self: TransformParentUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.transformId), self.transformId);
-        try ipc.write(@TypeOf(self.newParentId), self.newParentId);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TransformParentUpdate {
-        var self: TransformParentUpdate = undefined;
-        self.transformId = try ipc.read(@TypeOf(self.transformId));
-        self.newParentId = try ipc.read(@TypeOf(self.newParentId));
-        return self;
-    }
-};
-
-pub const TransformPoseUpdate = extern struct {
-    transformId: i32,
-    pose: RenderTransform,
-
-    pub fn write(self: TransformPoseUpdate, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.transformId), self.transformId);
-        try ipc.write(@TypeOf(self.pose), self.pose);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !TransformPoseUpdate {
-        var self: TransformPoseUpdate = undefined;
-        self.transformId = try ipc.read(@TypeOf(self.transformId));
-        self.pose = try ipc.read(@TypeOf(self.pose));
-        return self;
-    }
-};
-
-pub const RenderIntRect = extern struct {
-    x: i32,
-    y: i32,
-    width: i32,
-    height: i32,
-
-    pub fn write(self: RenderIntRect, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.x), self.x);
-        try ipc.write(@TypeOf(self.y), self.y);
-        try ipc.write(@TypeOf(self.width), self.width);
-        try ipc.write(@TypeOf(self.height), self.height);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderIntRect {
-        var self: RenderIntRect = undefined;
-        self.x = try ipc.read(@TypeOf(self.x));
-        self.y = try ipc.read(@TypeOf(self.y));
-        self.width = try ipc.read(@TypeOf(self.width));
-        self.height = try ipc.read(@TypeOf(self.height));
-        return self;
-    }
-};
-
-pub const RenderRect = extern struct {
-    x: f32,
-    y: f32,
-    width: f32,
-    height: f32,
-
-    pub fn write(self: RenderRect, ipc: IpcSerializer) !void {
-        try ipc.write(@TypeOf(self.x), self.x);
-        try ipc.write(@TypeOf(self.y), self.y);
-        try ipc.write(@TypeOf(self.width), self.width);
-        try ipc.write(@TypeOf(self.height), self.height);
-    }
-
-    pub fn read(ipc: IpcDeserializer) !RenderRect {
-        var self: RenderRect = undefined;
-        self.x = try ipc.read(@TypeOf(self.x));
-        self.y = try ipc.read(@TypeOf(self.y));
-        self.width = try ipc.read(@TypeOf(self.width));
-        self.height = try ipc.read(@TypeOf(self.height));
-        return self;
-    }
-};
+// NOTE: RendererCommand
